Fix long-appointment parts rendering, vertical Scheduler grouping (T714290) (#6987)
