+++ js/ui/scheduler/ui.scheduler.appointments.strategy.vertical.js
@@ -107,7 +107,7 @@ var VerticalRenderingStrategy = BaseAppointmentsStrategy.inherit({
         var tailHeight = appointmentGeometry.sourceAppointmentHeight - appointmentGeometry.reducedHeight,
             width = appointmentGeometry.width,
             result = [],
-            currentPartTop = 0,
+            currentPartTop = this.instance.fire("getGroupTop", appointmentSettings.groupIndex),
             offset = this.instance.fire("isGroupedByDate") ? this._defaultWidth * this.instance.fire("getGroupCount") : this._defaultWidth,
             left = appointmentSettings.left + offset;
 
@@ -117,6 +117,8 @@ var VerticalRenderingStrategy = BaseAppointmentsStrategy.inherit({
             if(tailHeight < minHeight) {
                 tailHeight = minHeight;
             }
+
+            currentPartTop += this.instance.fire("getOffsetByAllDayPanel", appointmentSettings.groupIndex);
 
             result.push(extend(true, {}, appointmentSettings, {
                 top: currentPartTop,
