+++ js/ui/scheduler/ui.scheduler.subscribes.js
@@ -466,6 +466,14 @@ var subscribes = {
         return result;
     },
 
+    getOffsetByAllDayPanel: function(groupIndex) {
+        return this._workSpace._getOffsetByAllDayPanel(groupIndex);
+    },
+
+    getGroupTop: function(groupIndex) {
+        return this._workSpace._getGroupTop(groupIndex);
+    },
+
     updateResizableArea: function() {
         var $allResizableElements = this.$element().find(".dx-scheduler-appointment.dx-resizable");
