+++ js/ui/scheduler/ui.scheduler.work_space.js
@@ -2090,11 +2090,23 @@ var SchedulerWorkSpace = Widget.inherit({
             rowIndex: position.rowIndex,
             cellIndex: position.cellIndex,
             hMax: horizontalHMax,
-            vMax: this._groupedStrategy.getVerticalMax(groupIndex),
+            vMax: this.getVerticalMax(groupIndex),
             groupIndex: groupIndex
         };
 
         return coordinates;
+    },
+
+    getVerticalMax: function(groupIndex) {
+        return this._groupedStrategy.getVerticalMax(groupIndex);
+    },
+
+    _getOffsetByAllDayPanel: function(groupIndex) {
+        return this._groupedStrategy._getOffsetByAllDayPanel(groupIndex);
+    },
+
+    _getGroupTop: function(groupIndex) {
+        return this._groupedStrategy._getGroupTop(groupIndex);
     },
 
     isGroupedByDate: function() {
