+++ testing/tests/DevExpress.ui.widgets.scheduler/integration.appointments.tests.js
@@ -26,7 +26,6 @@ var noop = require("core/utils/common").noop,
     dragEvents = require("events/drag"),
     DataSource = require("data/data_source/data_source").DataSource,
     CustomStore = require("data/custom_store"),
-    subscribes = require("ui/scheduler/ui.scheduler.subscribes"),
     dataUtils = require("core/element_data"),
     dateSerialization = require("core/utils/date_serialization");
 
@@ -39,10 +38,6 @@ var DATE_TABLE_CELL_CLASS = "dx-scheduler-date-table-cell",
 var APPOINTMENT_DEFAULT_OFFSET = 25,
     APPOINTMENT_MOBILE_OFFSET = 50;
 
-function getDeltaTz(schedulerTz, date) {
-    var defaultTz = date.getTimezoneOffset() * 60000;
-    return schedulerTz * 3600000 + defaultTz;
-}
 
 function getOffset() {
     if(devices.current().deviceType !== "desktop") {
@@ -455,341 +450,6 @@ QUnit.test("Tasks should be filtered by resources if groups are changed", functi
     assert.deepEqual(dataUtils.data($appointments.get(1), "dxItemData"), tasks[1], "The second appointment data is OK");
 });
 
-QUnit.test("Appointments on Day view should have a right height and position if startDate begins day before", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            startDate: new Date(2016, 9, 5, 23, 30),
-            endDate: new Date(2016, 9, 6, 1),
-            text: "new Date sample"
-        }],
-        currentDate: new Date(2016, 9, 6),
-        views: ["day"],
-        currentView: "day",
-        cellDuration: 60
-    });
-    var $element = this.instance.$element(),
-        $appointment = $element.find("." + APPOINTMENT_CLASS),
-        cellHeight = $element.find("." + DATE_TABLE_CELL_CLASS).first().get(0).getBoundingClientRect().height;
-
-    assert.equal($appointment.position().top, 0, "Appointment has a right top position");
-    assert.equal($appointment.outerHeight(), cellHeight, "Appointment has a right height");
-});
-
-QUnit.test("Appointments on Day view should have a right position if widget is small", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            startDate: new Date(2016, 9, 6, 1),
-            endDate: new Date(2016, 9, 6, 3),
-            text: "new Date sample"
-        }],
-        currentDate: new Date(2016, 9, 6),
-        views: ["week"],
-        width: 350,
-        currentView: "week",
-        cellDuration: 60
-    });
-    var $element = this.instance.$element(),
-        $appointment = $element.find("." + APPOINTMENT_CLASS),
-        cellWidth = $element.find("." + DATE_TABLE_CELL_CLASS).first().outerWidth();
-
-    assert.roughEqual($appointment.position().left, cellWidth * 4 + 50, 2, "Appointment has a right left position");
-});
-
-QUnit.test("Appointment with resources should have a right height and position if it ends on the next day", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            startDate: new Date(2017, 6, 22, 20),
-            endDate: new Date(2017, 6, 23, 4),
-            text: "Appointment",
-            ownerId: 1
-        }],
-        groups: ["ownerId"],
-        resources: [
-            {
-                field: "ownerId",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        currentDate: new Date(2017, 6, 22),
-        views: ["day"],
-        currentView: "day",
-        cellDuration: 60
-    });
-
-    var $element = this.instance.$element(),
-        $appointment = $element.find("." + APPOINTMENT_CLASS),
-        cellHeight = $element.find("." + DATE_TABLE_CELL_CLASS).first().get(0).getBoundingClientRect().height;
-
-    assert.equal($appointment.length, 1, "Only one appt is rendered");
-    assert.equal($appointment.position().top, cellHeight * 20, "Appointment has a right top position");
-    assert.equal($appointment.outerHeight(), cellHeight * 4, "Appointment has a right height");
-});
-
-QUnit.test("The part of the appointment that ends after midnight should be shown on Week view", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            startDate: new Date(2015, 4, 25, 22, 0),
-            endDate: new Date(2015, 4, 26, 2, 15),
-            text: "Test task"
-        }],
-        currentDate: new Date(2015, 4, 25),
-        views: ["week"],
-        currentView: "week",
-        cellDuration: 60
-    });
-
-    var $element = this.instance.$element(),
-        $appointment = $element.find("." + APPOINTMENT_CLASS);
-
-    assert.equal($appointment.length, 2, "ok");
-    assert.equal($appointment.eq(1).position().top, 0, "Appointment has a right top position");
-});
-
-QUnit.test("The part of the appointment that ends after midnight should have right height when set startDayHour & endDayHour", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            startDate: new Date(2015, 4, 25, 21),
-            endDate: new Date(2015, 4, 26, 2),
-            text: "Test task"
-        }],
-        currentDate: new Date(2015, 4, 25),
-        views: ["week"],
-        currentView: "week",
-        startDayHour: 1,
-        endDayHour: 23,
-        cellDuration: 60
-    });
-
-    var $element = this.instance.$element(),
-        $appointment = $element.find("." + APPOINTMENT_CLASS).eq(1),
-        cellHeight = $element.find("." + DATE_TABLE_CELL_CLASS).eq(0).get(0).getBoundingClientRect().height;
-
-    assert.equal($appointment.outerHeight(), cellHeight, "appt part has right height");
-});
-
-QUnit.test("The part of recurrence appointment after midnight should be shown on Day view", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            startDate: new Date(2015, 4, 25, 23, 0),
-            endDate: new Date(2015, 4, 26, 1, 15),
-            text: "Test task",
-            recurrenceRule: "FREQ=DAILY;INTERVAL=2"
-        }],
-        currentDate: new Date(2015, 4, 26),
-        views: ["day"],
-        currentView: "day",
-        cellDuration: 60
-    });
-
-    var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS);
-
-    assert.equal($appointment.length, 1, "Appt part is shown on 2d day");
-    assert.equal($appointment.eq(0).position().top, 0, "Appointment has a right top position");
-});
-
-QUnit.test("The part of recurrence appointment after midnight should have right height on the first day of week", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            startDate: new Date(2015, 4, 25, 22, 0),
-            endDate: new Date(2015, 4, 26, 3, 30),
-            text: "Test task",
-            recurrenceRule: "FREQ=DAILY;INTERVAL=2"
-        }],
-        currentDate: new Date(2015, 5, 1),
-        views: ["week"],
-        currentView: "week",
-        firstDayOfWeek: 1,
-        startDayHour: 2,
-        endDayHour: 23,
-        cellDuration: 30
-    });
-
-    var $element = this.instance.$element(),
-        $appointment = $element.find("." + APPOINTMENT_CLASS).eq(0),
-        cellHeight = $element.find("." + DATE_TABLE_CELL_CLASS).eq(0).get(0).getBoundingClientRect().height;
-
-    assert.equal($appointment.outerHeight(), cellHeight * 3, "appt part has right height");
-});
-
-QUnit.test("Appts should be filtered correctly with custom timeZone", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            startDate: new Date(Date.UTC(2016, 4, 4, 15)),
-            endDate: new Date(Date.UTC(2016, 4, 7, 15, 30)),
-            text: "new Date sample"
-        }, {
-            startDate: new Date(Date.UTC(2016, 4, 7, 23)),
-            endDate: new Date(Date.UTC(2016, 4, 7, 21, 59)),
-            text: "The last one"
-        }],
-        currentDate: new Date(Date.UTC(2016, 4, 7)),
-        timeZone: 5,
-        currentView: "day",
-        firstDayOfWeek: 1
-    });
-
-    assert.equal(this.instance.$element().find("." + APPOINTMENT_CLASS).length, 1, "Only one appt is rendered");
-});
-
-QUnit.test("Appts should be filtered correctly with custom timeZone as string", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            startDate: new Date(Date.UTC(2016, 4, 4, 15)),
-            endDate: new Date(Date.UTC(2016, 4, 7, 15, 30)),
-            text: "new Date sample"
-        }, {
-            startDate: new Date(Date.UTC(2016, 4, 7, 23)),
-            endDate: new Date(Date.UTC(2016, 4, 7, 21, 59)),
-            text: "The last one"
-        }],
-        currentDate: new Date(Date.UTC(2016, 4, 7)),
-        timeZone: "Asia/Calcutta",
-        currentView: "day",
-        firstDayOfWeek: 1
-    });
-
-    assert.equal(this.instance.$element().find("." + APPOINTMENT_CLASS).length, 1, "Only one appt is rendered");
-});
-
-QUnit.test("Appts should be filtered correctly if there is a custom tz and start day hour is not 0", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            startDate: "2015-05-27T23:00:00+01:00",
-            endDate: "2015-05-28T00:00:00+01:00",
-            text: "a"
-        }, {
-            startDate: "2015-05-26T18:30:00+01:00",
-            endDate: "2015-05-26T20:30:00+01:00",
-            text: "b"
-        }],
-        startDayHour: 7,
-        currentDate: new Date(2015, 4, 25),
-        timeZone: -8,
-        height: 500,
-        currentView: "week",
-        firstDayOfWeek: 1
-    });
-
-    var $element = this.instance.$element(),
-        $appt = $element.find("." + APPOINTMENT_CLASS),
-        cellHeight = $element.find("." + DATE_TABLE_CELL_CLASS).eq(0).get(0).getBoundingClientRect().height,
-        apptPosition = translator.locate($appt.eq(0)),
-        clientTzOffset = new Date("2015-05-27T23:00:00+01:00").getTimezoneOffset() / 60;
-
-    var cellsCount = (new Date(
-        new Date("2015-05-27T23:00:00+01:00").setHours(
-            new Date("2015-05-27T23:00:00+01:00").getHours() + clientTzOffset - 8)).getHours() - 7) * 2;
-
-    assert.equal($appt.length, 2, "Appts are OK");
-    assert.roughEqual(apptPosition.top, cellHeight * cellsCount, 2.001, "Appt top offset is OK");
-
-    this.instance.option({
-        currentView: "day",
-        currentDate: new Date(2015, 4, 26)
-    });
-
-    $appt = $element.find("." + APPOINTMENT_CLASS);
-    assert.equal($appt.length, 1, "Appts are OK on the Day view");
-});
-
-QUnit.test("Appts should be filtered correctly if there is a custom tz and start day hour is not 0(T396719)", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-
-    try {
-        this.createInstance({
-            dataSource: [{
-                text: "Stand-up meeting",
-                startDate: new Date(2015, 4, 25, 17),
-                endDate: new Date(2015, 4, 25, 17, 30),
-                startDateTimeZone: "America/Lima", // -5
-                endDateTimeZone: "America/Lima"
-            }],
-            startDayHour: 10,
-            currentDate: new Date(2015, 4, 25),
-            timeZone: "America/Dawson_Creek", // -7
-            height: 500,
-            currentView: "week",
-            firstDayOfWeek: 1
-        });
-
-        var apptCount = this.instance.$element().find("." + APPOINTMENT_CLASS).length;
-
-        assert.equal(apptCount, 0, "There are not appts");
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Recurring appointments should be rendered correctly with a custom timezone(T385377)", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-
-    try {
-        this.createInstance({
-            dataSource: [],
-            currentDate: new Date(2016, 4, 7),
-            timeZone: "Asia/Ashkhabad",
-            height: 500,
-            currentView: "week",
-            firstDayOfWeek: 1
-        });
-
-        this.instance.addAppointment({
-            startDate: new Date(2016, 4, 2),
-            endDate: new Date(2016, 4, 2, 0, 30),
-            recurrenceRule: "FREQ=DAILY"
-        });
-
-        var $element = this.instance.$element(),
-            $appt = this.instance.$element().find("." + APPOINTMENT_CLASS).eq(0),
-            expectedLeftPosition = $element.find(".dx-scheduler-time-panel").outerWidth(),
-            apptPosition = $appt.position();
-
-        assert.roughEqual(apptPosition.top, 0, 2.001, "Appts top is OK");
-        assert.roughEqual(apptPosition.left, expectedLeftPosition, 2.001, "Appts left is OK");
-
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Appointments should have correctly height with a custom timezone(T387561)", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-
-    try {
-        this.createInstance({
-            dataSource: [{
-                text: "Stand-up meeting",
-                startDate: "2015-05-25T17:00:00.000Z",
-                endDate: "2015-05-25T17:15:00.000Z",
-                startDateTimeZone: "Europe/Belgrade",
-                endDateTimeZone: "Europe/Belgrade"
-            }, {
-                text: "Approve New Online Marketing Strategy",
-                startDate: "2015-05-25T17:00:00.000Z",
-                endDate: "2015-05-25T19:00:00.000Z"
-            }],
-            currentDate: new Date(2015, 4, 25),
-            timeZone: "America/Los_Angeles",
-            height: 500,
-            currentView: "week",
-            firstDayOfWeek: 0
-        });
-
-        var $element = this.instance.$element(),
-            $appts = this.instance.$element().find("." + APPOINTMENT_CLASS),
-            cellHeight = $element.find("." + DATE_TABLE_CELL_CLASS).eq(0).outerHeight();
-
-        assert.roughEqual($appts.eq(0).outerHeight(), cellHeight / 2, 2.001, "Appts top is OK");
-        assert.roughEqual($appts.eq(1).outerHeight(), cellHeight * 4, 2.001, "Appts top is OK");
-
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
 QUnit.test("Scheduler tasks should have a right height", function(assert) {
     this.createInstance({ dataSource: this.tasks, currentDate: new Date(2015, 1, 9) });
     this.clock.tick();
@@ -816,31 +476,6 @@ QUnit.test("Scheduler tasks should have a right dimensions for month view", func
 
     assert.roughEqual($appointment.height(), $cell.outerHeight() * 0.6 / 2, 2, "Task has a right height");
     assert.roughEqual($appointment.outerWidth(), $cell.outerWidth(), 1.001, "Task has a right width");
-});
-
-QUnit.test("Recurring appointment icon should be visible on the month view", function(assert) {
-    this.createInstance({
-        dataSource: [
-            {
-                text: "Appt",
-                startDate: new Date(2015, 1, 9, 1),
-                endDate: new Date(2015, 1, 9, 10),
-                recurrenceRule: "FREQ=DAILY"
-            }
-        ],
-        currentDate: new Date(2015, 1, 9),
-        timeZone: "America/Los_Angeles",
-        views: ["month"],
-        currentView: "month",
-        height: 800
-    });
-
-    this.clock.tick();
-
-    var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(0),
-        $recurringIcon = $appointment.find(".dx-scheduler-appointment-recurrence-icon");
-
-    assert.ok($recurringIcon.parent().hasClass("dx-scheduler-appointment-content"), "Recurring icon is visible");
 });
 
 QUnit.test("Scheduler tasks should have a right height when currentView is changed", function(assert) {
@@ -881,49 +516,6 @@ QUnit.test("Two not rival appointments with fractional coordinates should have c
     assert.equal($appointment.length, 3, "All appointments are rendered");
     assert.equal(translator.locate($appointment.eq(0)).top, translator.locate($appointment.eq(1)).top, "appointment is rendered in right place");
     assert.equal(translator.locate($appointment.eq(1)).top, translator.locate($appointment.eq(2)).top, "appointment is rendered in right place");
-});
-
-QUnit.test("Two vertical neighbor appointments should be placed correctly", function(assert) {
-    this.createInstance({
-        dataSource: [],
-        firstDayOfWeek: 1,
-        currentDate: new Date(2015, 1, 24),
-        views: ["week"],
-        currentView: "week",
-        width: 800
-    });
-
-    this.instance.addAppointment({
-        text: "b",
-        startDate: new Date(2015, 1, 23, 1),
-        endDate: new Date(2015, 1, 23, 1, 30)
-    });
-
-    this.instance.addAppointment({
-        text: "a",
-        startDate: new Date(2015, 1, 23, 0, 30),
-        endDate: new Date(2015, 1, 23, 1)
-    });
-
-    this.instance.addAppointment({
-        text: "c",
-        startDate: new Date(2015, 1, 23, 0, 30),
-        endDate: new Date(2015, 1, 23, 1),
-        allDay: true
-    });
-
-    var $commonAppointments = this.instance.$element().find(".dx-scheduler-scrollable-appointments .dx-scheduler-appointment"),
-        $allDayAppts = this.instance.$element().find(".dx-scheduler-all-day-appointment"),
-        cellWidth = this.instance.$element().find("." + DATE_TABLE_CELL_CLASS).eq(0).outerWidth(),
-        appointmentOffset = getOffset();
-
-    assert.roughEqual(translator.locate($commonAppointments.eq(0)).left, 100, 2.001, "Left position is OK");
-    assert.roughEqual(translator.locate($commonAppointments.eq(1)).left, 100, 2.001, "Left position is OK");
-    assert.roughEqual($allDayAppts.eq(0).position().left, 100, 2.001, "Left position is OK");
-
-    assert.roughEqual($commonAppointments.eq(0).outerWidth(), cellWidth - appointmentOffset, 1.001, "Width is OK");
-    assert.roughEqual($commonAppointments.eq(1).outerWidth(), cellWidth - appointmentOffset, 1.001, "Width is OK");
-    assert.roughEqual($allDayAppts.eq(0).outerWidth(), cellWidth, 1.001, "Width is OK");
 });
 
 QUnit.test("DblClick on appointment should call scheduler.showAppointmentPopup", function(assert) {
@@ -990,6 +582,40 @@ QUnit.test("DblClick on appointment should not affect the related cell start dat
     } finally {
         this.instance.showAppointmentPopup.restore();
     }
+});
+
+QUnit.test("Recurrence repeat-end editor should be closed after reopening appointment popup", function(assert) {
+    this.createInstance({
+        currentDate: new Date(2015, 1, 9),
+        dataSource: new DataSource({
+            store: []
+        }),
+        currentView: "week"
+    });
+
+    var firstAppointment = { startDate: new Date(2015, 1, 9), endDate: new Date(2015, 1, 9, 1), text: "caption 1" },
+        secondAppointment = { startDate: new Date(2015, 1, 9), endDate: new Date(2015, 1, 9, 1), text: "caption 2" };
+
+    this.instance.showAppointmentPopup(firstAppointment);
+
+    var form = this.instance.getAppointmentDetailsForm(),
+        repeatOnEditor = form.$element().find(".dx-switch").eq(1).dxSwitch("instance"),
+        repeatEndEditor = form.getEditor("recurrenceRule")._switchEndEditor;
+
+    repeatOnEditor.option("value", true);
+
+    repeatEndEditor.option("value", true);
+    $(".dx-scheduler-appointment-popup").find(".dx-popup-done").trigger("dxclick");
+
+    this.instance.showAppointmentPopup(secondAppointment);
+
+    form = this.instance.getAppointmentDetailsForm(),
+    repeatOnEditor = form.$element().find(".dx-switch").eq(1).dxSwitch("instance"),
+    repeatEndEditor = form.getEditor("recurrenceRule")._switchEndEditor;
+
+    repeatOnEditor.option("value", true);
+
+    assert.notOk(repeatEndEditor.option("value"), "Switch is closed");
 });
 
 QUnit.test("Disabled appointment could not be focused", function(assert) {
@@ -1100,140 +726,6 @@ QUnit.test("Appointment should be copied before sending to the details view", fu
         formData = detailsForm.option("formData");
 
     assert.notEqual(formData, task, "Appointment data is copied");
-});
-
-QUnit.test("Appointment startDate and endDate should be correct in the details view, if custom timeZone is setting", function(assert) {
-    var startDate = new Date(2015, 3, 11, 11),
-        endDate = new Date(2015, 3, 11, 11, 30);
-
-    var task = {
-        text: "Task 1",
-        Start: startDate,
-        End: endDate
-    };
-
-    var timezone = 5;
-
-    this.createInstance({
-        dataSource: new DataSource({
-            store: [task]
-        }),
-        currentDate: new Date(2015, 3, 23),
-        startDateExpr: "Start",
-        endDateExpr: "End",
-        timeZone: timezone
-    });
-
-
-    this.instance.showAppointmentPopup(task);
-
-    var detailsForm = this.instance.getAppointmentDetailsForm(),
-        formData = detailsForm.option("formData"),
-        deltaTz = getDeltaTz(timezone, startDate);
-
-    assert.deepEqual(formData.Start, new Date(startDate.getTime() + deltaTz), "start date is correct");
-    assert.deepEqual(formData.End, new Date(endDate.getTime() + deltaTz), "end date is correct");
-});
-
-QUnit.test("Appointment startDate and endDate should be correct in the details view, if custom timeZone is setting as string", function(assert) {
-    var startDate = new Date(2015, 3, 11, 11),
-        endDate = new Date(2015, 3, 11, 11, 30);
-
-    var task = {
-        text: "Task 1",
-        Start: startDate,
-        End: endDate
-    };
-
-    this.clock.restore();
-
-    this.createInstance({
-        dataSource: new DataSource({
-            store: [task]
-        }),
-        currentDate: new Date(2015, 3, 23),
-        startDateExpr: "Start",
-        endDateExpr: "End",
-        timeZone: "Asia/Ashkhabad"
-    });
-
-    this.instance.showAppointmentPopup(task);
-
-    var detailsForm = this.instance.getAppointmentDetailsForm(),
-        formData = detailsForm.option("formData"),
-        deltaTz = getDeltaTz(5, startDate);
-
-    assert.deepEqual(formData.Start, new Date(startDate.getTime() + deltaTz), "start date is correct");
-    assert.deepEqual(formData.End, new Date(endDate.getTime() + deltaTz), "end date is correct");
-});
-
-QUnit.test("Appointment startDate and endDate should be correct in the details view for new appointment, if custom timeZone was set", function(assert) {
-    this.createInstance({
-        dataSource: new DataSource({
-            store: []
-        }),
-        currentDate: new Date(2015, 3, 23),
-        startDateExpr: "Start",
-        endDateExpr: "End",
-        timeZone: "Asia/Calcutta"
-    });
-
-    pointerMock(this.instance.$element().find("." + DATE_TABLE_CELL_CLASS).eq(22)).start().click().click();
-
-    var detailsForm = this.instance.getAppointmentDetailsForm(),
-        formData = detailsForm.option("formData");
-
-    assert.deepEqual(formData.Start, new Date(2015, 3, 23, 11), "start date is correct");
-    assert.deepEqual(formData.End, new Date(2015, 3, 23, 11, 30), "end date is correct");
-});
-
-QUnit.test("Appointments should have correct size with custom time zone & hourly bounds", function(assert) {
-
-    this.clock.restore();
-
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-
-    try {
-        this.createInstance({
-            currentDate: new Date(2015, 4, 25),
-            startDayHour: 8,
-            endDayHour: 18,
-            views: ["day"],
-            currentView: "day",
-            firstDayOfWeek: 1,
-            dataSource: [{
-                text: "Approve New Online Marketing Strategy",
-                startDate: new Date(2015, 4, 25, 8),
-                endDate: new Date(2015, 4, 25, 12),
-                startDateTimeZone: "Africa/Brazzaville",
-                endDateTimeZone: "Africa/Brazzaville"
-            }, {
-                text: "Stand-up meeting",
-                startDate: new Date(2015, 4, 25, 18),
-                endDate: new Date(2015, 4, 25, 20),
-                startDateTimeZone: "Africa/Brazzaville",
-                endDateTimeZone: "Africa/Brazzaville"
-            }],
-            timeZone: "Africa/Brazzaville" // +1
-        });
-
-        var $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS),
-            $first = $appointments.eq(0),
-            $second = $appointments.eq(1),
-            cellHeight = this.instance.$element().find("." + DATE_TABLE_CELL_CLASS).eq(0).outerHeight();
-
-        assert.roughEqual($first.outerHeight(), cellHeight * 4, 2.001, "Appointment height is correct");
-        assert.roughEqual($second.outerHeight(), cellHeight * 4, 2.001, "Appointment height is correct");
-
-        assert.equal($first.find(".dx-scheduler-appointment-content-date").eq(0).text(), "6:00 AM", "Appointment start is correct");
-        assert.equal($first.find(".dx-scheduler-appointment-content-date").eq(2).text(), "10:00 AM", "Appointment edn is correct");
-
-        assert.equal($second.find(".dx-scheduler-appointment-content-date").eq(0).text(), "4:00 PM", "Appointment start is correct");
-        assert.equal($second.find(".dx-scheduler-appointment-content-date").eq(2).text(), "6:00 PM", "Appointment edn is correct");
-
-    } finally {
-        tzOffsetStub.restore();
-    }
 });
 
 QUnit.test("Add new appointment", function(assert) {
@@ -1402,38 +894,6 @@ QUnit.test("Scheduler should not update scroll position if appointment is visibl
     }
 });
 
-QUnit.test("Scheduler should not update scroll position if appointment is visible, when timeZone is set ", function(assert) {
-    var workSpace;
-
-    try {
-        this.clock.restore();
-
-        this.createInstance({
-            startDayHour: 3,
-            endDayHour: 10,
-            currentDate: new Date(Date.UTC(2015, 1, 9)).toJSON(),
-            dataSource: [],
-            currentView: "week",
-            height: 500,
-            timeZone: "Asia/Ashkhabad"
-        });
-
-        this.instance.getWorkSpaceScrollable().scrollBy(170);
-
-        workSpace = this.instance.getWorkSpace();
-
-        var appointment = { startDate: new Date(Date.UTC(2015, 1, 9, 3)).toJSON(), endDate: new Date(Date.UTC(2015, 1, 9, 3, 30)).toJSON(), text: "caption" },
-            scrollToTimeSpy = sinon.spy(workSpace, "scrollToTime");
-
-        this.instance.showAppointmentPopup(appointment);
-        $(".dx-scheduler-appointment-popup .dx-popup-done").trigger("dxclick");
-
-        assert.notOk(scrollToTimeSpy.calledOnce, "scrollToTime was not called");
-    } finally {
-        workSpace.scrollToTime.restore();
-    }
-});
-
 QUnit.test("Scheduler should update scroll position if appointment was added to invisible bottom area", function(assert) {
     this.createInstance({
         currentDate: new Date(2015, 1, 9),
@@ -1442,31 +902,6 @@ QUnit.test("Scheduler should update scroll position if appointment was added to 
         }),
         currentView: "week",
         height: 300
-    });
-
-    var appointment = { startDate: new Date(2015, 1, 9, 21), endDate: new Date(2015, 1, 9, 22), text: "caption 2" },
-        workSpace = this.instance.getWorkSpace(),
-        scrollToTimeSpy = sinon.spy(workSpace, "scrollToTime");
-
-    try {
-        this.instance.showAppointmentPopup(appointment);
-        $(".dx-scheduler-appointment-popup .dx-popup-done").trigger("dxclick");
-
-        assert.ok(scrollToTimeSpy.calledOnce, "scrollToTime was called");
-    } finally {
-        workSpace.scrollToTime.restore();
-    }
-});
-
-QUnit.test("Scheduler should update scroll position if appointment was added to invisible bottom area, timezone is set", function(assert) {
-    this.createInstance({
-        currentDate: new Date(2015, 1, 9),
-        dataSource: new DataSource({
-            store: []
-        }),
-        currentView: "week",
-        height: 300,
-        timezone: "Asia/Ashkhabad"
     });
 
     var appointment = { startDate: new Date(2015, 1, 9, 21), endDate: new Date(2015, 1, 9, 22), text: "caption 2" },
@@ -1640,166 +1075,6 @@ QUnit.test("Update appointment", function(assert) {
     }
 });
 
-QUnit.test("Appointment date correction should be rollback after closing popup, if custom timeZone was set", function(assert) {
-    var updatedItem = {
-            text: "Task 1",
-            startDate: new Date(2015, 1, 7, 1),
-            endDate: new Date(2015, 1, 7, 2)
-        },
-        data = new DataSource({
-            store: [updatedItem]
-        });
-
-    this.createInstance({
-        currentView: "week",
-        currentDate: new Date(2015, 1, 7),
-        dataSource: data,
-        timeZone: 5
-    });
-
-    var updateAppointment = sinon.spy(this.instance, "updateAppointment");
-    try {
-        this.instance.showAppointmentPopup(updatedItem);
-
-        $(".dx-scheduler-appointment-popup .dx-popup-done").trigger("dxclick");
-
-        tooltip.hide();
-
-        assert.ok(updateAppointment.calledOnce, "Update method is called");
-        assert.deepEqual(updateAppointment.getCall(0).args[0], updatedItem, "Target item is correct");
-        assert.deepEqual(updateAppointment.getCall(0).args[1], updatedItem, "New data is correct");
-    } finally {
-        updateAppointment.restore();
-    }
-});
-
-QUnit.test("Appointment date correction should be rollback after closing popup, if custom timeZone was set as string", function(assert) {
-    var updatedItem = {
-            text: "Task 1",
-            startDate: new Date(2015, 1, 7, 1),
-            endDate: new Date(2015, 1, 7, 2)
-        },
-        data = new DataSource({
-            store: [updatedItem]
-        });
-
-    this.createInstance({
-        currentView: "week",
-        currentDate: new Date(2015, 1, 7),
-        dataSource: data,
-        timeZone: "Asia/Calcutta"
-    });
-
-    var updateAppointment = sinon.spy(this.instance, "updateAppointment");
-
-    try {
-        this.instance.showAppointmentPopup(updatedItem);
-
-        $(".dx-scheduler-appointment-popup .dx-popup-done").trigger("dxclick");
-
-        assert.ok(updateAppointment.calledOnce, "Update method is called");
-        assert.deepEqual(updateAppointment.getCall(0).args[0], updatedItem, "Target item is correct");
-        assert.deepEqual(updateAppointment.getCall(0).args[1], updatedItem, "New data is correct");
-    } finally {
-        updateAppointment.restore();
-    }
-});
-
-QUnit.test("Appointment should have a correct template with custom timezone", function(assert) {
-    this.clock.restore();
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(new Date(2016, 4, 7, 5).getTimezoneOffset() * 60000);
-
-    try {
-        this.createInstance({
-            currentDate: new Date(2016, 4, 7),
-            startDayHour: 7,
-            views: ["day"],
-            currentView: "day",
-            dataSource: [
-                {
-                    startDate: new Date(Date.UTC(2016, 4, 7, 5)),
-                    startDateTimeZone: "Asia/Qyzylorda", // +6:00
-                    endDateTimeZone: "Asia/Qyzylorda",
-                    endDate: new Date(Date.UTC(2016, 4, 7, 5, 30)),
-                    text: 'new Date sample'
-                }
-            ],
-            timeZone: "Asia/Ashkhabad"// +5:00
-        });
-
-        var $appt = this.instance.$element().find("." + APPOINTMENT_CLASS),
-            $contentDates = $appt.find(".dx-scheduler-appointment-content-date");
-
-        assert.equal($contentDates.first().text(), "10:00 AM", "Start date is correct");
-        assert.equal($contentDates.last().text(), "10:30 AM", "End date is correct");
-
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("onAppointmentAdding event args should be consistent with adding appointment when custom timezone (T686572)", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-
-    try {
-        this.createInstance({
-            currentDate: new Date(2016, 4, 7),
-            dateSerializationFormat: "yyyy-MM-ddTHH:mm:ssZ",
-            timeZone: 'Etc/UTC',
-            views: ["day"],
-            currentView: "day",
-            dataSource: [],
-            height: 800,
-            onAppointmentAdding: function(e) {
-                assert.equal(e.appointmentData.startDate, "2016-05-07T08:00:00Z", "Start date is ok");
-                assert.equal(e.appointmentData.endDate, "2016-05-07T08:30:00Z", "End date is ok");
-            }
-        });
-
-        this.instance.addAppointment({
-            startDate: new Date(Date.UTC(2016, 4, 7, 5)),
-            endDate: new Date(Date.UTC(2016, 4, 7, 5, 30)),
-            startDateTimeZone: "Asia/Qyzylorda", // +6:00
-            endDateTimeZone: "Asia/Qyzylorda",
-            text: 'new Date sample'
-        });
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Appointment should have a correct template with custom timezone(T387040)", function(assert) {
-    this.clock.restore();
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(new Date(2016, 4, 7, 5).getTimezoneOffset() * 60000);
-
-    try {
-        this.createInstance({
-            currentDate: new Date(2016, 4, 7),
-            startDayHour: 7,
-            views: ["day"],
-            currentView: "day",
-            dataSource: []
-        });
-
-        this.instance.option("dataSource", [{
-            startDate: new Date(Date.UTC(2016, 4, 7, 5)),
-            startDateTimeZone: "Asia/Qyzylorda",
-            endDateTimeZone: "Asia/Qyzylorda",
-            endDate: new Date(Date.UTC(2016, 4, 7, 5, 30)),
-            text: 'new Date sample'
-        }]);
-
-        var $appt = this.instance.$element().find("." + APPOINTMENT_CLASS),
-            $contentDates = $appt.find(".dx-scheduler-appointment-content-date");
-
-        assert.equal($contentDates.first().text(), "11:00 AM", "Start date is correct");
-        assert.equal($contentDates.last().text(), "11:30 AM", "End date is correct");
-
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
 QUnit.test("updateAppointment method should be called when task was resized", function(assert) {
     var data = new DataSource({
         store: this.tasks
@@ -1920,477 +1195,6 @@ QUnit.test("Non-grid-aligned appointments should be resized correctly, when endD
     assert.deepEqual(this.instance.option("dataSource")[0].endDate, new Date(2015, 1, 9, 15), "End date is OK");
 });
 
-QUnit.test("Appointment with custom timezone should be resized correctly(T390801)", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-
-    try {
-        this.createInstance({
-            currentDate: new Date(2015, 5, 12),
-            views: ["week"],
-            currentView: "week",
-            editing: true,
-            timeZone: "America/Araguaina", // -3
-            dataSource: [{
-                text: "a",
-                startDate: new Date(2015, 5, 12, 10),
-                endDate: new Date(2015, 5, 12, 12)
-            }]
-        });
-
-        var cellHeight = this.instance.$element().find("." + DATE_TABLE_CELL_CLASS).eq(0).get(0).getBoundingClientRect().height,
-            $appointment = this.instance.$element().find("." + APPOINTMENT_CLASS),
-            initialAppointmentTop = $appointment.position().top;
-
-        var pointer = pointerMock(this.instance.$element().find(".dx-resizable-handle-bottom").eq(0)).start();
-
-        pointer.dragStart().drag(0, cellHeight);
-        pointer.dragEnd();
-
-        $appointment = this.instance.$element().find("." + APPOINTMENT_CLASS).eq(0);
-
-        assert.equal($appointment.position().top, initialAppointmentTop, "Appointment top is OK");
-        assert.roughEqual($appointment.outerHeight(), cellHeight * 5, 2.001, "Appointment height is OK");
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Recurrence appointment with custom timezone should be resized correctly(T390801)", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-
-    try {
-        this.createInstance({
-            currentDate: new Date(2015, 5, 12),
-            views: ["week"],
-            currentView: "week",
-            editing: true,
-            recurrenceEditMode: "occurrence",
-            timeZone: "America/Araguaina", // -3
-            dataSource: [{
-                text: "a",
-                startDate: new Date(2015, 5, 12, 10).toString(),
-                endDate: new Date(2015, 5, 12, 12).toString(),
-                recurrenceRule: "FREQ=DAILY"
-            }]
-        });
-
-        var cellHeight = this.instance.$element().find("." + DATE_TABLE_CELL_CLASS).eq(0).get(0).getBoundingClientRect().height,
-            $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS),
-            initialAppointmentTop = $appointments.eq(0).position().top;
-
-        var pointer = pointerMock(this.instance.$element().find(".dx-resizable-handle-bottom").eq(0)).start();
-
-        pointer.dragStart().drag(0, cellHeight);
-        pointer.dragEnd();
-
-        $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS);
-
-        assert.equal($appointments.length, 2, "Appointment count is OK");
-        assert.equal($appointments.eq(1).position().top, initialAppointmentTop, "Appointment top is OK");
-        assert.roughEqual($appointments.eq(1).outerHeight(), cellHeight * 5, 2.001, "Appointment height is OK");
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Recurrence appointment with custom tz that isn't equal to scheduler tz should be resized correctly(T390801)", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-    try {
-        this.createInstance({
-            currentDate: new Date(2015, 5, 12),
-            views: ["week"],
-            currentView: "week",
-            editing: true,
-            recurrenceEditMode: "occurrence",
-            timeZone: "America/Araguaina", // -3
-            dataSource: [{
-                text: "a",
-                startDate: new Date(2015, 5, 12, 10).toString(),
-                endDate: new Date(2015, 5, 12, 12).toString(),
-                recurrenceRule: "FREQ=DAILY",
-                startDateTimeZone: "America/Lima", // -5
-                endDateTimeZone: "America/Lima"
-            }]
-        });
-
-        var cellHeight = this.instance.$element().find("." + DATE_TABLE_CELL_CLASS).eq(0).get(0).getBoundingClientRect().height,
-            $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS),
-            initialAppointmentTop = $appointments.eq(0).position().top;
-
-        var pointer = pointerMock(this.instance.$element().find(".dx-resizable-handle-bottom").eq(0)).start();
-
-        pointer.dragStart().drag(0, cellHeight);
-        pointer.dragEnd();
-
-        $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS);
-
-        assert.equal($appointments.length, 2, "Appointment count is OK");
-        assert.equal($appointments.eq(1).position().top, initialAppointmentTop, "Appointment top is OK");
-        assert.roughEqual($appointments.eq(1).outerHeight(), cellHeight * 5, 2.001, "Appointment height is OK");
-
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Recurrence appointment with custom tz that isn't equal to scheduler tz should be opened correctly(T390801)", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-    try {
-        var appointment = {
-            text: "Stand-up meeting",
-            startDate: new Date(2016, 5, 7, 9),
-            endDate: new Date(2016, 5, 7, 10),
-            startDateTimeZone: "Europe/Moscow", // +3
-            endDateTimeZone: "Europe/Moscow",
-            recurrenceRule: "FREQ=DAILY"
-        };
-
-        var initialPosition;
-
-        this.createInstance({
-            currentDate: new Date(2016, 5, 7),
-            views: ["week"],
-            currentView: "week",
-            firstDayOfWeek: 1,
-            recurrenceEditMode: "occurrence",
-            timeZone: "America/Phoenix", // -7
-            dataSource: [appointment]
-        });
-
-        var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(1);
-
-        $appointment.trigger("dxdblclick");
-        initialPosition = $appointment.position();
-
-        $(".dx-scheduler-appointment-popup .dx-popup-done").trigger("dxclick");
-        var updatedPosition = this.instance.$element().find("." + APPOINTMENT_CLASS).not(".dx-scheduler-appointment-recurrence").position();
-
-        assert.equal(updatedPosition.top, initialPosition.top, "Top is updated correctly");
-        assert.equal(updatedPosition.left, initialPosition.left, "Left is updated correctly");
-
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Arguments in event args should be correct when timezone is set(T579457)", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-    try {
-        var appointment = {
-            startDate: new Date('2017-11-22T14:30:00.000Z'),
-            endDate: new Date('2017-11-22T15:00:00.000Z'),
-            allDay: false,
-            recurrenceRule: "FREQ=DAILY;COUNT=3",
-            text: ""
-        };
-
-        this.createInstance({
-            currentDate: new Date(2017, 10, 22),
-            views: ["week"],
-            currentView: "week",
-            firstDayOfWeek: 1,
-            onAppointmentClick: function(args) {
-                assert.equal(args.appointmentData.startDate.getTime(), args.targetedAppointmentData.startDate.getTime(), "Arguments are OK");
-                assert.equal(args.appointmentData.endDate.getTime(), args.targetedAppointmentData.endDate.getTime(), "Arguments are OK");
-            },
-            timeZone: 'Etc/UTC',
-            dataSource: [appointment]
-        });
-
-        var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(0);
-        $appointment.trigger("dxclick");
-
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Recurrence appointment with the same custom timezones should be opened correctly(T390801)", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-    try {
-        var appointment = {
-            text: "Stand-up meeting",
-            startDate: "2015-05-25T15:30:00.000Z",
-            endDate: "2015-05-25T15:45:00.000Z",
-            startDateTimeZone: "America/Phoenix", // -7
-            endDateTimeZone: "America/Phoenix",
-            recurrenceRule: "FREQ=DAILY"
-        };
-
-        var initialPosition;
-
-        this.createInstance({
-            currentDate: new Date(2015, 4, 25),
-            views: ["week"],
-            currentView: "week",
-            firstDayOfWeek: 1,
-            recurrenceEditMode: "occurrence",
-            timeZone: "America/Phoenix", // -7
-            dataSource: [appointment]
-        });
-
-        var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(1);
-
-        $appointment.trigger("dxdblclick");
-        initialPosition = $appointment.position();
-
-        $(".dx-scheduler-appointment-popup .dx-popup-done").trigger("dxclick");
-        var updatedPosition = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).not(".dx-scheduler-appointment-recurrence").position();
-
-        assert.equal(updatedPosition.top, initialPosition.top, "Top is updated correctly");
-        assert.equal(updatedPosition.left, initialPosition.left, "Left is updated correctly");
-
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Appointment with custom tz that isn't equal to scheduler tz should be resized correctly(T392414)", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-
-    try {
-        this.createInstance({
-            currentDate: new Date(2015, 4, 25),
-            views: ["week"],
-            currentView: "week",
-            editing: true,
-            recurrenceEditMode: "occurrence",
-            timeZone: "America/Phoenix", // -7
-            dataSource: [{
-                text: "a",
-                startDate: "2015-05-25T17:00:00.000Z",
-                endDate: "2015-05-25T17:15:00.000Z",
-                startDateTimeZone: "America/Lima", // -5
-                endDateTimeZone: "America/Lima"
-            }]
-        });
-
-        var cellHeight = this.instance.$element().find("." + DATE_TABLE_CELL_CLASS).eq(0).outerHeight(),
-            $appointment = this.instance.$element().find("." + APPOINTMENT_CLASS).first(),
-            initialAppointmentTop = $appointment.position().top;
-
-        var pointer = pointerMock(this.instance.$element().find(".dx-resizable-handle-bottom").eq(0)).start();
-
-        pointer.dragStart().drag(0, cellHeight);
-        pointer.dragEnd();
-
-        $appointment = this.instance.$element().find("." + APPOINTMENT_CLASS).first();
-
-        assert.equal($appointment.position().top, initialAppointmentTop, "Appointment top is OK");
-        assert.roughEqual($appointment.outerHeight(), cellHeight * 2, 2.001, "Appointment height is OK");
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Appointment with custom tz that is equal to scheduler tz should be resized correctly(T392414)", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-
-    try {
-        this.createInstance({
-            currentDate: new Date(2015, 4, 25),
-            views: ["week"],
-            currentView: "week",
-            editing: true,
-            recurrenceEditMode: "occurrence",
-            timeZone: "America/Lima", // -5
-            dataSource: [{
-                text: "a",
-                startDate: "2015-05-25T17:00:00.000Z",
-                endDate: "2015-05-25T17:15:00.000Z",
-                startDateTimeZone: "America/Lima",
-                endDateTimeZone: "America/Lima"
-            }]
-        });
-
-        var cellHeight = this.instance.$element().find("." + DATE_TABLE_CELL_CLASS).eq(0).outerHeight(),
-            $appointment = this.instance.$element().find("." + APPOINTMENT_CLASS).first(),
-            initialAppointmentTop = $appointment.position().top;
-
-        var pointer = pointerMock(this.instance.$element().find(".dx-resizable-handle-bottom").eq(0)).start();
-        pointer.dragStart().drag(0, cellHeight);
-        pointer.dragEnd();
-
-        $appointment = this.instance.$element().find("." + APPOINTMENT_CLASS).first();
-
-        assert.equal($appointment.position().top, initialAppointmentTop, "Appointment top is OK");
-        assert.roughEqual($appointment.outerHeight(), cellHeight * 2, 2.001, "Appointment height is OK");
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Appointment with custom tz should be resized correctly if the scheduler tz is empty(T392414)", function(assert) {
-    this.createInstance({
-        currentDate: new Date(2015, 11, 25),
-        startDayHour: 6,
-        views: ["day"],
-        currentView: "day",
-        editing: true,
-        dataSource: [{
-            text: "a",
-            startDate: "2015-12-25T17:00:00.000Z",
-            endDate: "2015-12-25T17:15:00.000Z",
-            startDateTimeZone: "America/Lima", // -5
-            endDateTimeZone: "America/Lima"
-        }]
-    });
-
-    var cellHeight = this.instance.$element().find("." + DATE_TABLE_CELL_CLASS).eq(0).outerHeight(),
-        $appointment = this.instance.$element().find("." + APPOINTMENT_CLASS).first(),
-        initialAppointmentTop = $appointment.position().top;
-
-    var pointer = pointerMock(this.instance.$element().find(".dx-resizable-handle-bottom").eq(0)).start();
-    pointer.dragStart().drag(0, cellHeight);
-    pointer.dragEnd();
-
-    $appointment = this.instance.$element().find("." + APPOINTMENT_CLASS).first();
-
-    assert.equal($appointment.position().top, initialAppointmentTop, "Appointment top is OK");
-    assert.roughEqual($appointment.outerHeight(), cellHeight * 2, 2.001, "Appointment height is OK");
-});
-
-QUnit.test("Appointment with custom tz that isn't equal to scheduler tz should be dragged correctly(T392414)", function(assert) {
-    this.createInstance({
-        currentDate: new Date(2015, 4, 25),
-        startDayHour: 6,
-        views: ["day"],
-        currentView: "day",
-        editing: true,
-        recurrenceEditMode: "occurrence",
-        timeZone: "America/Phoenix", // -7
-        dataSource: [{
-            text: "a",
-            startDate: "2015-05-25T17:00:00.000Z",
-            endDate: "2015-05-25T17:15:00.000Z",
-            startDateTimeZone: "America/Lima", // -5
-            endDateTimeZone: "America/Lima"
-        }]
-    });
-
-    var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).first(),
-        $cell = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).eq(6),
-        initialAppointmentHeight = $appointment.outerHeight();
-
-    $appointment.trigger(dragEvents.start);
-    $cell.trigger(dragEvents.enter);
-    $appointment.trigger(dragEvents.end);
-
-    $appointment = this.instance.$element().find("." + APPOINTMENT_CLASS).first();
-
-
-    assert.roughEqual($appointment.position().top, $cell.outerHeight() * 6, 2.001, "Appointment top is OK");
-    assert.equal($appointment.outerHeight(), initialAppointmentHeight, "Appointment height is OK");
-
-    var startDateText = $appointment.find(".dx-scheduler-appointment-content-date").eq(0).text(),
-        endDateText = $appointment.find(".dx-scheduler-appointment-content-date").eq(2).text(),
-        cellData = dataUtils.data($cell.get(0), "dxCellData"),
-        startDate = cellData.startDate,
-        endDate = new Date(cellData.startDate.getTime() + 15 * 60 * 1000);
-
-    assert.equal(startDateText, dateLocalization.format(startDate, "shorttime"), "Appointment start date is OK");
-    assert.equal(endDateText, dateLocalization.format(endDate, "shorttime"), "Appointment end date is OK");
-});
-
-QUnit.test("Appointment with 'Etc/UTC' tz should be rendered correctly(T394991)", function(assert) {
-    var tzOffsetStub;
-    try {
-        this.clock.restore();
-        tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(new Date("2016-06-25T17:00:00.000Z").getTimezoneOffset() * 60000);
-
-        this.createInstance({
-            currentDate: new Date(2016, 5, 25),
-            startDayHour: 16,
-            views: ["day"],
-            currentView: "day",
-            editing: true,
-            timeZone: "Greenwich", // 0
-            dataSource: [{
-                text: "a",
-                startDate: "2016-06-25T17:00:00.000Z",
-                endDate: "2016-06-25T17:30:00.000Z"
-            }]
-        });
-
-        var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).first(),
-            $cell = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).eq(6),
-            initialAppointmentHeight = $appointment.outerHeight();
-
-        assert.roughEqual($appointment.position().top, $cell.outerHeight() * 2, 2.001, "Appointment top is OK");
-        assert.roughEqual($appointment.outerHeight(), $cell.outerHeight(), 2.001, "Appointment height is OK");
-
-        $appointment.trigger(dragEvents.start);
-        $cell.trigger(dragEvents.enter);
-        $appointment.trigger(dragEvents.end);
-
-        $appointment = this.instance.$element().find("." + APPOINTMENT_CLASS).first();
-
-        assert.roughEqual($appointment.position().top, $cell.outerHeight() * 6, 2.001, "Appointment top is OK");
-        assert.equal($appointment.outerHeight(), initialAppointmentHeight, "Appointment height is OK");
-
-        var startDateText = $appointment.find(".dx-scheduler-appointment-content-date").eq(0).text(),
-            endDateText = $appointment.find(".dx-scheduler-appointment-content-date").eq(2).text(),
-            cellData = dataUtils.data($cell.get(0), "dxCellData"),
-            startDate = cellData.startDate,
-            endDate = new Date(cellData.startDate.getTime() + 30 * 60 * 1000);
-
-        assert.equal(startDateText, dateLocalization.format(startDate, "shorttime"), "Appointment start date is OK");
-        assert.equal(endDateText, dateLocalization.format(endDate, "shorttime"), "Appointment end date is OK");
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Recurrence appointment with 'Etc/UTC' tz should be updated correctly via drag(T394991)", function(assert) {
-    var tzOffsetStub;
-    try {
-        this.clock.restore();
-        tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(new Date("2015-12-25T17:00:00.000Z").getTimezoneOffset() * 60000);
-
-        this.createInstance({
-            currentDate: new Date(2015, 11, 25),
-            startDayHour: 16,
-            views: ["week"],
-            currentView: "week",
-            editing: true,
-            timeZone: "Etc/UTC", // 0
-            recurrenceEditMode: "occurrence",
-            firstDayOfWeek: 1,
-            dataSource: [{
-                text: "a",
-                startDate: "2015-12-25T17:00:00.000Z",
-                endDate: "2015-12-25T17:30:00.000Z",
-                recurrenceRule: "FREQ=DAILY"
-            }]
-        });
-
-        var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).first(),
-            $cell = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).eq(21),
-            initialAppointmentHeight = $appointment.outerHeight();
-
-        $appointment.trigger(dragEvents.start);
-        $cell.trigger(dragEvents.enter);
-        $appointment.trigger(dragEvents.end);
-
-        $appointment = this.instance.$element().find("." + APPOINTMENT_CLASS).not(".dx-scheduler-appointment-recurrence");
-
-        assert.roughEqual($appointment.position().top, $cell.outerHeight() * 3, 2.001, "Appointment top is OK");
-        assert.equal($appointment.outerHeight(), initialAppointmentHeight, "Appointment height is OK");
-
-
-        var startDateText = $appointment.find(".dx-scheduler-appointment-content-date").eq(0).text(),
-            endDateText = $appointment.find(".dx-scheduler-appointment-content-date").eq(2).text(),
-            cellData = dataUtils.data($cell.get(0), "dxCellData"),
-            startDate = cellData.startDate,
-            endDate = new Date(cellData.startDate.getTime() + 30 * 60 * 1000);
-
-        assert.equal(startDateText, dateLocalization.format(startDate, "shorttime"), "Appointment start date is OK");
-        assert.equal(endDateText, dateLocalization.format(endDate, "shorttime"), "Appointment end date is OK");
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
 // TODO: also need test when task is dragging outside the area. updated dates should be equal to old dates
 QUnit.test("Task dragging", function(assert) {
     var data = new DataSource({
@@ -2417,46 +1221,6 @@ QUnit.test("Task dragging", function(assert) {
     this.clock.tick();
     assert.equal(dataSourceItem.text, updatedItem.text, "New data is correct");
     assert.equal(dataSourceItem.allDay, updatedItem.allDay, "New data is correct");
-    assert.deepEqual(dataSourceItem.startDate, updatedItem.startDate, "New data is correct");
-    assert.deepEqual(dataSourceItem.endDate, updatedItem.endDate, "New data is correct");
-});
-
-QUnit.test("Task dragging when custom timeZone is set", function(assert) {
-    this.clock.restore();
-    var timezone = -5,
-        timezoneDifference = getDeltaTz(timezone, new Date(2015, 1, 9)),
-        startDate = new Date(new Date(2015, 1, 9).getTime() - timezoneDifference),
-        endDate = new Date(new Date(2015, 1, 9, 1).getTime() - timezoneDifference);
-
-    var data = new DataSource({
-        store: [
-            {
-                text: "Task 1",
-                startDate: startDate,
-                endDate: endDate
-            }
-        ]
-    });
-
-    this.createInstance({ currentDate: new Date(2015, 1, 9), dataSource: data, editing: true, timeZone: timezone });
-
-    this.clock.tick();
-
-    var hour = 3600000;
-    var updatedItem = {
-        text: "Task 1",
-        startDate: new Date(startDate.getTime() + hour),
-        endDate: new Date(endDate.getTime() + hour),
-        allDay: false
-    };
-
-    $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(0).trigger(dragEvents.start);
-    $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).eq(2).trigger(dragEvents.enter);
-    $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(0).trigger(dragEvents.end);
-
-    var dataSourceItem = this.instance.option("dataSource").items()[0];
-
-    this.clock.tick();
     assert.deepEqual(dataSourceItem.startDate, updatedItem.startDate, "New data is correct");
     assert.deepEqual(dataSourceItem.endDate, updatedItem.endDate, "New data is correct");
 });
@@ -3612,56 +2376,6 @@ QUnit.test("Month appointment inside grouped view should have a right resizable 
     assert.equal($appointment.dxResizable("instance").option("area").right, initialResizableAreaRight - scrollOffset);
 });
 
-QUnit.test("Appointment size should depend on neighbor appointments", function(assert) {
-    var items = [{ text: "a", startDate: new Date(2015, 2, 4, 1), endDate: new Date(2015, 2, 4, 2) },
-        { text: "b", startDate: new Date(2015, 2, 4, 2, 30), endDate: new Date(2015, 2, 4, 3) },
-        { text: "c", startDate: new Date(2015, 2, 4, 2, 30), endDate: new Date(2015, 2, 4, 3) },
-        { text: "d", startDate: new Date(2015, 2, 4, 1, 30), endDate: new Date(2015, 2, 4, 3) }];
-
-    this.createInstance({
-        currentView: "week",
-        currentDate: new Date(2015, 2, 4),
-        dataSource: items
-    });
-
-    var $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS);
-    assert.roughEqual($appointments.eq(0).width(), $appointments.eq(3).width(), 0.1);
-});
-
-QUnit.test("Recurrence repeat-end editor should be closed after reopening appointment popup", function(assert) {
-    this.createInstance({
-        currentDate: new Date(2015, 1, 9),
-        dataSource: new DataSource({
-            store: []
-        }),
-        currentView: "week"
-    });
-
-    var firstAppointment = { startDate: new Date(2015, 1, 9), endDate: new Date(2015, 1, 9, 1), text: "caption 1" },
-        secondAppointment = { startDate: new Date(2015, 1, 9), endDate: new Date(2015, 1, 9, 1), text: "caption 2" };
-
-    this.instance.showAppointmentPopup(firstAppointment);
-
-    var form = this.instance.getAppointmentDetailsForm(),
-        repeatOnEditor = form.$element().find(".dx-switch").eq(1).dxSwitch("instance"),
-        repeatEndEditor = form.getEditor("recurrenceRule")._switchEndEditor;
-
-    repeatOnEditor.option("value", true);
-
-    repeatEndEditor.option("value", true);
-    $(".dx-scheduler-appointment-popup").find(".dx-popup-done").trigger("dxclick");
-
-    this.instance.showAppointmentPopup(secondAppointment);
-
-    form = this.instance.getAppointmentDetailsForm(),
-    repeatOnEditor = form.$element().find(".dx-switch").eq(1).dxSwitch("instance"),
-    repeatEndEditor = form.getEditor("recurrenceRule")._switchEndEditor;
-
-    repeatOnEditor.option("value", true);
-
-    assert.notOk(repeatEndEditor.option("value"), "Switch is closed");
-});
-
 QUnit.test("Rival appointments should have correct positions on month view, rtl mode", function(assert) {
     this.createInstance({
         rtlEnabled: true,
@@ -4154,73 +2868,6 @@ QUnit.test("Recurrence long appointment should be rendered correctly when curren
     assert.equal($appointment.length, 10, "Appointments were rendered");
 });
 
-QUnit.test("Recurrence appointment should be rendered correctly when currentDate was changed: all-day", function(assert) {
-    var appointment = {
-        startDate: new Date(2015, 1, 4, 0),
-        endDate: new Date(2015, 1, 4, 0, 30),
-        text: "long appointment",
-        recurrenceRule: "FREQ=DAILY",
-        allDay: true
-    };
-
-    this.createInstance({
-        currentDate: new Date(2015, 1, 4),
-        dataSource: [appointment],
-        views: ["day"],
-        currentView: "day"
-    });
-
-    this.instance.option("currentDate", new Date(2015, 1, 5));
-
-    var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS);
-
-    assert.equal($appointment.length, 1, "Appointment is rendered");
-});
-
-QUnit.test("Recurrence appointment should be rendered correctly when currentDate was changed: day", function(assert) {
-    var appointment = {
-        startDate: new Date(2015, 1, 4, 0),
-        endDate: new Date(2015, 1, 4, 1),
-        text: "long appointment",
-        recurrenceRule: "FREQ=DAILY"
-    };
-
-    this.createInstance({
-        currentDate: new Date(2015, 1, 4),
-        dataSource: [appointment],
-        views: ["day"],
-        currentView: "day"
-    });
-
-    this.instance.option("currentDate", new Date(2015, 1, 5));
-
-    var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS);
-
-    assert.roughEqual($appointment.height(), 100, 1.001, "Appointment is rendered correctly");
-});
-
-QUnit.test("Appointments should have correct position, rtl mode, editing=false", function(assert) {
-    var appointment = {
-        startDate: new Date(2015, 1, 4, 0),
-        endDate: new Date(2015, 1, 4, 1)
-    };
-
-    this.createInstance({
-        rtlEnabled: true,
-        editing: false,
-        currentDate: new Date(2015, 1, 4),
-        views: ["day"],
-        currentView: "day",
-        firstDayOfWeek: 1,
-        dataSource: [appointment]
-    });
-
-    var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(0),
-        appointmentOffset = getOffset();
-
-    assert.roughEqual($appointment.position().left, appointmentOffset, 2, "Appointment left is correct on init");
-});
-
 QUnit.test("Appointment should be rendered correctly with expressions on init", function(assert) {
     var startDate = new Date(2015, 1, 4, 0),
         endDate = new Date(2015, 1, 4, 1);
@@ -4284,249 +2931,6 @@ QUnit.test("Appointment should be rendered correctly with recurrenceRule express
     assert.equal($recAppointment.find(".dx-scheduler-appointment-content-date").eq(0).text(), dateLocalization.format(startDate, "shorttime"), "Start Date is correct on init");
     assert.equal($recAppointment.find(".dx-scheduler-appointment-content-date").eq(2).text(), dateLocalization.format(endDate, "shorttime"), "End Date is correct on init");
     assert.equal($recAppointment.find(".dx-scheduler-appointment-recurrence-icon").length, 1, "Recurrence icon is rendered");
-});
-
-QUnit.test("Appointment should have correct height, when startDayHour is decimal", function(assert) {
-    var appointments = [{
-        startDate: new Date(2015, 1, 4, 5, 35).toString(),
-        endDate: new Date(2015, 1, 4, 5, 45).toString(),
-        text: "abc"
-    }];
-
-    this.createInstance({
-        currentDate: new Date(2015, 1, 4),
-        cellDuration: 5,
-        views: ["day"],
-        currentView: "day",
-        firstDayOfWeek: 1,
-        dataSource: appointments,
-        startDayHour: 5.5
-    });
-
-    var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(0),
-        cellHeight = this.instance.$element().find("." + DATE_TABLE_CELL_CLASS).eq(0).get(0).getBoundingClientRect().height;
-
-    assert.roughEqual($appointment.position().top, cellHeight, 2.001, "Appointment top is correct");
-    assert.roughEqual($appointment.outerHeight(), 2 * cellHeight, 2.001, "Appointment height is correct");
-});
-
-QUnit.test("Appointment should be rendered correctly when custom timezone was set", function(assert) {
-    var startDate = new Date(2015, 1, 4, 5),
-        endDate = new Date(2015, 1, 4, 7);
-
-    var appointments = [{
-        startDate: startDate.toString(),
-        endDate: endDate.toString(),
-        text: "abc",
-        recurrenceRule: "FREQ=DAILY"
-    }];
-
-    var timezone = 5;
-
-    this.createInstance({
-        currentDate: new Date(2015, 1, 4),
-        views: ["day"],
-        currentView: "day",
-        firstDayOfWeek: 1,
-        dataSource: appointments,
-        timeZone: timezone
-    });
-
-    var $recAppointment = this.instance.$element().find("." + APPOINTMENT_CLASS).eq(0);
-
-    assert.equal($recAppointment.find(".dx-scheduler-appointment-content div").eq(0).text(), "abc", "Text is correct on init");
-
-    var deltaTz = getDeltaTz(timezone, startDate);
-    assert.equal($recAppointment.find(".dx-scheduler-appointment-content-date").eq(0).text(), dateLocalization.format(new Date(startDate.getTime() + deltaTz), "shorttime"), "Start Date is correct on init");
-    assert.equal($recAppointment.find(".dx-scheduler-appointment-content-date").eq(2).text(), dateLocalization.format(new Date(endDate.getTime() + deltaTz), "shorttime"), "End Date is correct on init");
-    assert.equal($recAppointment.find(".dx-scheduler-appointment-recurrence-icon").length, 1, "Recurrence icon is rendered");
-});
-
-QUnit.test("Appointment should be rendered correctly when custom timezone was set as string", function(assert) {
-    var startDate = new Date(2015, 1, 4, 5),
-        endDate = new Date(2015, 1, 4, 7);
-
-    var appointments = [{
-        startDate: startDate.toString(),
-        endDate: endDate.toString(),
-        text: "abc",
-        recurrenceRule: "FREQ=DAILY"
-    }
-    ];
-
-    this.clock.restore();
-
-    this.createInstance({
-        currentDate: new Date(2015, 1, 4),
-        views: ["day"],
-        currentView: "day",
-        firstDayOfWeek: 1,
-        dataSource: appointments,
-        timeZone: "Asia/Ashkhabad"
-    });
-
-
-    var $recAppointment = this.instance.$element().find("." + APPOINTMENT_CLASS).eq(0);
-
-    assert.equal($recAppointment.find(".dx-scheduler-appointment-content div").eq(0).text(), "abc", "Text is correct on init");
-
-    var deltaTz = getDeltaTz(5, startDate);
-    assert.equal($recAppointment.find(".dx-scheduler-appointment-content-date").eq(0).text(), dateLocalization.format(new Date(startDate.getTime() + deltaTz), "shorttime"), "Start Date is correct on init");
-    assert.equal($recAppointment.find(".dx-scheduler-appointment-content-date").eq(2).text(), dateLocalization.format(new Date(endDate.getTime() + deltaTz), "shorttime"), "End Date is correct on init");
-    assert.equal($recAppointment.find(".dx-scheduler-appointment-recurrence-icon").length, 1, "Recurrence icon is rendered");
-});
-
-QUnit.test("Appointment should be rendered correctly when appointment timeZone was set", function(assert) {
-    var appointments = [{
-        startDate: new Date(2015, 1, 4, 5).toString(),
-        startDateTimeZone: "Asia/Calcutta", // +05:30
-        endDateTimeZone: "Asia/Calcutta",
-        endDate: new Date(2015, 1, 4, 6).toString(),
-        text: "abc"
-    }];
-
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-
-    try {
-        this.createInstance({
-            currentDate: new Date(2015, 1, 4),
-            views: ["day"],
-            currentView: "day",
-            firstDayOfWeek: 1,
-            dataSource: appointments
-        });
-
-        var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(0);
-
-        assert.equal($appointment.find(".dx-scheduler-appointment-content-date").eq(0).text(), dateLocalization.format(new Date(2015, 1, 4, 7, 30), "shorttime"), "Start Date is correct on init");
-        assert.equal($appointment.find(".dx-scheduler-appointment-content-date").eq(2).text(), dateLocalization.format(new Date(2015, 1, 4, 8, 30), "shorttime"), "End Date is correct on init");
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Appointment should have correct height, when appointment timeZones were set", function(assert) {
-
-    this.clock.restore();
-
-    var appointments = [{
-        startDate: new Date(2015, 1, 4, 5).toString(),
-        startDateTimeZone: "Europe/Moscow",
-        endDateTimeZone: "Asia/Ashkhabad",
-        endDate: new Date(2015, 1, 4, 6).toString(),
-        text: "abc"
-    }];
-
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-
-    try {
-        this.createInstance({
-            currentDate: new Date(2015, 1, 4),
-            views: ["day"],
-            currentView: "day",
-            firstDayOfWeek: 1,
-            dataSource: appointments,
-            startDayHour: 5
-        });
-
-        var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(0),
-            cellHeight = this.instance.$element().find("." + DATE_TABLE_CELL_CLASS).eq(0).get(0).getBoundingClientRect().height;
-
-        assert.roughEqual($appointment.position().top, 0, 2.001, "Appointment top is correct");
-        assert.roughEqual($appointment.outerHeight(), 6 * cellHeight, 2.001, "Appointment height is correct");
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
-QUnit.test("Appointment should be rendered correctly when appointment timezone and scheduler timezone was set", function(assert) {
-    var startDate = new Date(2015, 1, 4, 5),
-        endDate = new Date(2015, 1, 4, 6);
-
-    var appointments = [{
-        startDate: startDate.toString(),
-        endDate: endDate.toString(),
-        startDateTimezone: "Asia/Ashkhabad",
-        text: "abc"
-    }];
-
-    this.clock.restore();
-
-    this.createInstance({
-        currentDate: new Date(2015, 1, 4),
-        views: ["day"],
-        currentView: "day",
-        firstDayOfWeek: 1,
-        dataSource: appointments,
-        timeZone: "Asia/Qyzylorda"
-    });
-
-    var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(0);
-
-    var deltaTz = getDeltaTz(6, startDate);
-
-    assert.equal($appointment.find(".dx-scheduler-appointment-content div").eq(0).text(), "abc", "Text is correct on init");
-
-    assert.equal($appointment.find(".dx-scheduler-appointment-content-date").eq(0).text(), dateLocalization.format(new Date(startDate.getTime() + deltaTz), "shorttime"), "Start Date is correct on init");
-    assert.equal($appointment.find(".dx-scheduler-appointment-content-date").eq(2).text(), dateLocalization.format(new Date(endDate.getTime() + deltaTz), "shorttime"), "End Date is correct on init");
-});
-
-QUnit.test("All-day Appointment should be rendered correctly when custom timezone was set", function(assert) {
-    var timezone = 5,
-        timezoneDifference = getDeltaTz(timezone, new Date(2016, 4, 4)),
-        startDate = new Date(new Date(2016, 4, 4).getTime() - timezoneDifference),
-        endDate = new Date(new Date(2016, 4, 5).getTime() - timezoneDifference);
-
-    this.createInstance({
-        views: ["week"],
-        currentView: "week",
-        currentDate: new Date(2016, 4, 3),
-        timeZone: "Asia/Ashkhabad",
-        dataSource: [{
-            startDate: startDate,
-            endDate: endDate
-        }]
-    });
-
-    var $element = this.instance.$element(),
-        apptWidth = $element.find("." + APPOINTMENT_CLASS).first().outerWidth(),
-        cellWidth = $element.find(".dx-scheduler-all-day-table-cell").first().outerWidth();
-
-    assert.roughEqual(apptWidth, cellWidth, 2.001, "Appt width is OK");
-});
-
-QUnit.test("Appointment should be rendered correctly if timeZones is changed", function(assert) {
-
-    this.clock.restore();
-
-    var appointments = [{
-        startDate: new Date(2015, 1, 4, 5).toString(),
-        endDate: new Date(2015, 1, 4, 6).toString(),
-        text: "abc"
-    }];
-
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-
-    try {
-        this.createInstance({
-            currentDate: new Date(2015, 1, 4),
-            views: ["day"],
-            currentView: "day",
-            firstDayOfWeek: 1,
-            dataSource: appointments,
-            startDayHour: 5,
-            timeZone: 3
-        });
-
-        this.instance.option("timeZone", 4);
-
-        var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(0),
-            cellHeight = this.instance.$element().find("." + DATE_TABLE_CELL_CLASS).eq(0).outerHeight();
-
-        assert.roughEqual($appointment.position().top, cellHeight * 2, 2.001, "Appointment top is correct");
-    } finally {
-        tzOffsetStub.restore();
-    }
 });
 
 QUnit.test("Appointment should be rendered correctly with expressions on optionChanged", function(assert) {
@@ -4633,65 +3037,6 @@ QUnit.test("DropDown appointment should be rendered correctly with expressions o
     var $appointment = $(".dx-dropdownmenu-list .dx-item").first();
 
     assert.equal($appointment.find(".custom-title").text(), "Item 2", "Text is correct on init");
-});
-
-QUnit.test("DropDown appointment should be rendered correctly when timezone is set", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-    try {
-        var data = [
-            {
-                schedule: "Appointment 1",
-                startDate: new Date(2018, 8, 17, 1),
-                endDate: new Date(2018, 8, 17, 2)
-            },
-            {
-                schedule: "Appointment 2",
-                startDate: new Date(2018, 8, 17, 1),
-                endDate: new Date(2018, 8, 17, 2)
-            },
-            {
-                schedule: "Appointment 3",
-                startDate: new Date(2018, 8, 17, 1),
-                endDate: new Date(2018, 8, 17, 2)
-            },
-            {
-                schedule: "Appointment 4",
-                startDate: new Date(2018, 8, 17, 1),
-                endDate: new Date(2018, 8, 17, 2)
-            },
-            {
-                schedule: "Appointment 5",
-                startDate: new Date(2018, 8, 17, 1),
-                endDate: new Date(2018, 8, 17, 2)
-            },
-            {
-                schedule: "Appointment 6",
-                startDate: new Date(2018, 8, 17, 1),
-                endDate: new Date(2018, 8, 17, 2)
-            }
-        ];
-
-        this.createInstance({
-            dataSource: data,
-            views: ["month"],
-            currentView: "month",
-            currentDate: new Date(2018, 8, 17),
-            timeZone: 'Etc/UTC',
-            showCurrentTimeIndicator: false,
-            maxAppointmentsPerCell: 1,
-            height: 600,
-            textExpr: "schedule"
-        });
-
-        $(this.instance.$element()).find(".dx-scheduler-dropdown-appointments").dxDropDownMenu("instance").open();
-
-        var $appointment = $(".dx-dropdownmenu-list .dx-item").first(),
-            $dates = $appointment.find(".dx-scheduler-dropdown-appointment-date").first();
-
-        assert.equal($dates.text(), "September 16, 10:00 PM - 11:00 PM", "Dates is correct");
-    } finally {
-        tzOffsetStub.restore();
-    }
 });
 
 QUnit.test("DropDown button should be rendered correctly when appointmentCollectorTemplate is used", function(assert) {
@@ -5165,54 +3510,6 @@ QUnit.test("DropDown appointment button should have correct width on timeline vi
     assert.roughEqual($dropDownButton.outerWidth(), cellWidth - 4, 1.5, "DropDown button has correct width");
 });
 
-QUnit.test("dropDown appointment should not compact class on vertical view", function(assert) {
-    this.createInstance({
-        currentDate: new Date(2015, 4, 25),
-        views: [{ type: "week", name: "week" }],
-        currentView: "week",
-        maxAppointmentsPerCell: 'auto'
-    });
-
-    this.instance.option("dataSource", [
-        { text: '1', startDate: new Date(2015, 4, 25), endDate: new Date(2015, 4, 25, 1) },
-        { text: '2', startDate: new Date(2015, 4, 25), endDate: new Date(2015, 4, 25, 1) },
-        { text: '3', startDate: new Date(2015, 4, 25), endDate: new Date(2015, 4, 25, 1) }
-    ]);
-
-    var $dropDown = $(this.instance.$element()).find(".dx-scheduler-dropdown-appointments").eq(0);
-
-    assert.ok($dropDown.hasClass("dx-scheduler-dropdown-appointments-compact"), "class is ok");
-});
-
-QUnit.test("Appointment should have right width in workspace with timezone", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-    try {
-        this.clock.restore();
-        this.createInstance({
-            dataSource: [],
-            currentDate: new Date(2017, 4, 1),
-            currentView: "month",
-            firstDayOfWeek: 1,
-            startDayHour: 3,
-            endDayHour: 24,
-            timeZone: "Asia/Ashkhabad"
-        });
-
-        this.instance.addAppointment({
-            text: "Task 1",
-            startDate: new Date(2017, 4, 4),
-            endDate: new Date(2017, 4, 5)
-        });
-
-        var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(0),
-            $cell = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).eq(9);
-
-        assert.roughEqual($appointment.outerWidth(), $cell.outerWidth(), 1.001, "Task has a right width");
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
 QUnit.test("Appointment with zero-duration should be rendered correctly(T443143)", function(assert) {
 
     this.createInstance({
@@ -5496,112 +3793,6 @@ QUnit.test("FormData should be reset on saveChanges, dateSerializationFormat is 
     assert.deepEqual(endDateFormat, "yyyy-MM-ddTHH:mm:ss", "Appointment EndDate format is OK");
 });
 
-QUnit.test("Appointments should be rendered correctly, Day view with intervalCount", function(assert) {
-    var tasks = [
-        { text: "One", startDate: new Date(2015, 2, 16, 7), endDate: new Date(2015, 2, 16, 7, 30) },
-        { text: "Two", startDate: new Date(2015, 2, 16, 11), endDate: new Date(2015, 2, 16, 11, 30) },
-        { text: "Three", startDate: new Date(2015, 2, 18, 12), endDate: new Date(2015, 2, 18, 12, 30) },
-        { text: "Four", startDate: new Date(2015, 2, 18, 15), endDate: new Date(2015, 2, 18, 15, 30) }
-    ];
-    var dataSource = new DataSource({
-        store: tasks
-    });
-    this.createInstance({
-        currentDate: new Date(2015, 2, 16),
-        dataSource: dataSource,
-        views: [{
-            type: "day",
-            name: "Day",
-            intervalCount: 3
-        }],
-        currentView: "day"
-    });
-
-    var $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS);
-
-    assert.equal($appointments.length, 4, "Appointments were rendered correctly");
-});
-
-QUnit.test("Appointments should be rendered correctly, Week view with intervalCount", function(assert) {
-    var tasks = [
-        { text: "One", startDate: new Date(2015, 4, 25, 7), endDate: new Date(2015, 4, 25, 7, 30) },
-        { text: "Two", startDate: new Date(2015, 5, 1, 11), endDate: new Date(2015, 5, 1, 11, 30) },
-        { text: "Three", startDate: new Date(2015, 5, 6, 12), endDate: new Date(2015, 5, 6, 12, 30) },
-        { text: "Four", startDate: new Date(2015, 5, 12, 15), endDate: new Date(2015, 5, 12, 15, 30) }
-    ];
-    var dataSource = new DataSource({
-        store: tasks
-    });
-    this.createInstance({
-        currentDate: new Date(2015, 4, 25),
-        dataSource: dataSource,
-        views: [{
-            type: "week",
-            name: "week",
-            intervalCount: 3
-        }],
-        currentView: "week"
-    });
-
-    var $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS);
-
-    assert.equal($appointments.length, 4, "Appointments were rendered correctly");
-});
-
-QUnit.test("Appointments should be rendered correctly, Day view with intervalCount and startDate", function(assert) {
-    var tasks = [
-        { text: "One", startDate: new Date(2017, 5, 25, 4), endDate: new Date(2017, 5, 25, 4, 30) },
-        { text: "Two", startDate: new Date(2017, 5, 26, 0), endDate: new Date(2017, 5, 26, 0, 30) },
-        { text: "Three", startDate: new Date(2017, 5, 27, 10), endDate: new Date(2017, 5, 27, 11) }
-    ];
-    var dataSource = new DataSource({
-        store: tasks
-    });
-    this.createInstance({
-        currentDate: new Date(2017, 5, 26),
-        dataSource: dataSource,
-        views: [{
-            type: "day",
-            name: "day",
-            intervalCount: 3,
-            startDate: new Date(2017, 5, 25)
-        }],
-        currentView: "day"
-    });
-
-    var $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS);
-
-    assert.equal($appointments.length, 3, "Appointments were rendered correctly");
-});
-
-QUnit.test("Appointments should be rendered correctly, Week view with intervalCount and startDate", function(assert) {
-    var tasks = [
-        { text: "One", startDate: new Date(2017, 5, 22, 4), endDate: new Date(2017, 5, 22, 4, 30) },
-        { text: "Two", startDate: new Date(2017, 5, 26, 0), endDate: new Date(2017, 5, 26, 0, 30) },
-        { text: "Three", startDate: new Date(2017, 6, 2, 10), endDate: new Date(2017, 6, 2, 11) },
-        { text: "Four", startDate: new Date(2017, 6, 9, 8), endDate: new Date(2017, 6, 9, 8, 30) }
-    ];
-    var dataSource = new DataSource({
-        store: tasks
-    });
-    this.createInstance({
-        currentDate: new Date(2017, 5, 26),
-        dataSource: dataSource,
-        views: [{
-            type: "week",
-            name: "week",
-            intervalCount: 3,
-            startDate: new Date(2017, 5, 19)
-        }],
-        currentView: "week",
-        firstDayOfWeek: 1
-    });
-
-    var $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS);
-
-    assert.equal($appointments.length, 4, "Appointments were rendered correctly");
-});
-
 QUnit.test("Appointments should be rendered correctly, Month view with intervalCount and startDate", function(assert) {
     var tasks = [
         { text: "One", startDate: new Date(2017, 5, 22, 4), endDate: new Date(2017, 5, 22, 4, 30) },
@@ -5624,63 +3815,6 @@ QUnit.test("Appointments should be rendered correctly, Month view with intervalC
         currentView: "month",
         firstDayOfWeek: 1
     });
-
-    var $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS);
-
-    assert.equal($appointments.length, 3, "Appointments were rendered correctly");
-});
-
-QUnit.test("Appointments should be rendered correctly after switching Day view with intervalCount and startDate", function(assert) {
-    var tasks = [
-        { text: "One", startDate: new Date(2017, 5, 28, 4), endDate: new Date(2017, 5, 28, 4, 30) },
-        { text: "Two", startDate: new Date(2017, 5, 29, 0), endDate: new Date(2017, 5, 29, 0, 30) },
-        { text: "Three", startDate: new Date(2017, 5, 30, 10), endDate: new Date(2017, 5, 30, 11) }
-    ];
-    var dataSource = new DataSource({
-        store: tasks
-    });
-    this.createInstance({
-        currentDate: new Date(2017, 5, 26),
-        dataSource: dataSource,
-        views: [{
-            type: "day",
-            name: "day",
-            intervalCount: 3,
-            startDate: new Date(2017, 5, 25)
-        }],
-        currentView: "day"
-    });
-
-    $(this.instance.$element().find(".dx-scheduler-navigator-next")).trigger("dxclick");
-
-    var $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS);
-
-    assert.equal($appointments.length, 3, "Appointments were rendered correctly");
-});
-
-QUnit.test("Appointments should be rendered correctly after switching Week view with intervalCount and startDate", function(assert) {
-    var tasks = [
-        { text: "One", startDate: new Date(2017, 6, 10, 4), endDate: new Date(2017, 6, 10, 4, 30) },
-        { text: "Two", startDate: new Date(2017, 6, 18, 0), endDate: new Date(2017, 6, 18, 0, 30) },
-        { text: "Three", startDate: new Date(2017, 6, 25, 10), endDate: new Date(2017, 6, 25, 11) }
-    ];
-    var dataSource = new DataSource({
-        store: tasks
-    });
-    this.createInstance({
-        currentDate: new Date(2017, 5, 26),
-        dataSource: dataSource,
-        views: [{
-            type: "week",
-            name: "week",
-            intervalCount: 3,
-            startDate: new Date(2017, 5, 19)
-        }],
-        currentView: "week",
-        firstDayOfWeek: 1
-    });
-
-    $(this.instance.$element().find(".dx-scheduler-navigator-next")).trigger("dxclick");
 
     var $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS);
 
@@ -5712,374 +3846,6 @@ QUnit.test("Scheduler should add only one appointment at multiple 'done' button 
     var $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS);
 
     assert.equal($appointments.length, 1, "right appointment quantity");
-});
-
-QUnit.test("Appointment should have right width on mobile devices & desktop in week view", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            text: "a",
-            startDate: new Date(2018, 2, 13, 1),
-            endDate: new Date(2018, 2, 13, 3)
-        }],
-        currentDate: new Date(2018, 2, 13),
-        views: ["week"],
-        currentView: "week"
-    });
-
-    var expectedOffset = getOffset(),
-        $appointments = this.instance.$element().find("." + APPOINTMENT_CLASS),
-        cellWidth = this.instance.$element().find("." + DATE_TABLE_CELL_CLASS).eq(0).outerWidth();
-
-    assert.roughEqual($appointments.eq(0).outerWidth(), cellWidth - expectedOffset, 1.001, "Width is OK");
-});
-
-QUnit.test("Appointments should be rendered correctly in vertical grouped workspace Day", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            text: "a",
-            startDate: new Date(2018, 2, 1, 9),
-            endDate: new Date(2018, 2, 1, 10, 30),
-            id: 1
-        }, {
-            text: "b",
-            startDate: new Date(2018, 2, 1, 9),
-            endDate: new Date(2018, 2, 1, 10, 30),
-            id: 2
-        }],
-        currentDate: new Date(2018, 2, 1),
-        views: [{
-            type: "day",
-            groupOrientation: "vertical"
-        }],
-        currentView: "day",
-        groups: ["id"],
-        resources: [
-            {
-                field: "id",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        startDayHour: 9,
-        showAllDayPanel: false
-    });
-
-    var $appointments = $(this.instance.$element()).find("." + APPOINTMENT_CLASS);
-    assert.equal($appointments.length, 2, "two appointments are rendered");
-
-    var cellHeight = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).first().get(0).getBoundingClientRect().height;
-
-    assert.equal($appointments.eq(0).position().top, 0, "correct top position");
-    assert.equal($appointments.eq(0).position().left, 200, "correct left position");
-    assert.equal($appointments.eq(1).position().top, cellHeight * 30, "correct top position");
-    assert.equal($appointments.eq(1).position().left, 200, "correct left position");
-});
-
-QUnit.test("Appointments should be rendered correctly in vertical grouped workspace Week", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            text: "a",
-            startDate: new Date(2018, 2, 16, 9),
-            endDate: new Date(2018, 2, 16, 10, 30),
-            id: 1
-        }, {
-            text: "b",
-            startDate: new Date(2018, 2, 16, 9),
-            endDate: new Date(2018, 2, 16, 10, 30),
-            id: 2
-        }],
-        currentDate: new Date(2018, 2, 16),
-        views: [{
-            type: "week",
-            groupOrientation: "vertical"
-        }],
-        currentView: "week",
-        groups: ["id"],
-        resources: [
-            {
-                field: "id",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        startDayHour: 9,
-        showAllDayPanel: false
-    });
-
-    var $appointments = $(this.instance.$element()).find("." + APPOINTMENT_CLASS);
-    assert.equal($appointments.length, 2, "two appointments is rendered");
-
-    var cellHeight = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).eq(0).get(0).getBoundingClientRect().height,
-        cellWidth = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).eq(0).get(0).getBoundingClientRect().width;
-
-    assert.equal($appointments.eq(0).position().top, 0, "correct top position");
-    assert.equal($appointments.eq(0).position().left, 200 + cellWidth * 5, "correct left position");
-    assert.equal($appointments.eq(1).position().top, cellHeight * 30, "correct top position");
-    assert.equal($appointments.eq(1).position().left, 200 + cellWidth * 5, "correct left position");
-});
-
-QUnit.test("Appointments should be rendered correctly in vertical grouped workspace Week, showAllDayPanel = true", function(assert) {
-    this.createInstance({
-        dataSource: [
-            {
-                text: "1",
-                id: 2,
-                startDate: new Date(2018, 4, 21, 9, 30),
-                endDate: new Date(2018, 4, 21, 11, 30)
-            },
-            {
-                text: "2",
-                id: 2,
-                allDay: true,
-                startDate: new Date(2018, 4, 21, 9, 30),
-                endDate: new Date(2018, 4, 21, 11, 30)
-            },
-        ],
-        views: [{
-            type: "week",
-            groupOrientation: "vertical"
-        }],
-        currentView: "week",
-        groups: ["id"],
-        resources: [
-            {
-                field: "id",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        currentDate: new Date(2018, 4, 21),
-        startDayHour: 9,
-        endDayHour: 15,
-        cellDuration: 60,
-        showAllDayPanel: true,
-        width: 2000,
-        maxAppointmentsPerCell: "auto"
-    });
-
-    var $appointments = $(this.instance.$element()).find("." + APPOINTMENT_CLASS);
-    assert.equal($appointments.length, 2, "two appointments are rendered");
-
-    var cellHeight = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).first().outerHeight();
-
-    assert.roughEqual($appointments.eq(0).position().top, 7 * cellHeight, 1.5, "correct top position of allDay appointment");
-    assert.roughEqual($appointments.eq(0).outerHeight(), 0.5 * cellHeight, 2, "correct size of allDay appointment");
-    assert.roughEqual($appointments.eq(0).position().left, 456, 1.1, "correct left position of allDay appointment");
-
-    assert.roughEqual($appointments.eq(1).position().top, 8.5 * cellHeight, 1.5, "correct top position of appointment");
-    assert.roughEqual($appointments.eq(1).position().left, 456, 1.1, "correct left position of appointment");
-});
-
-QUnit.test("Rival allDay appointments from different groups should be rendered correctly in vertical grouped workspace Week", function(assert) {
-    this.createInstance({
-        dataSource: [
-            {
-                text: "1",
-                id: 1,
-                allDay: true,
-                startDate: new Date(2018, 4, 21, 9, 30),
-                endDate: new Date(2018, 4, 21, 11, 30)
-            },
-            {
-                text: "2",
-                id: 2,
-                allDay: true,
-                startDate: new Date(2018, 4, 21, 9, 30),
-                endDate: new Date(2018, 4, 21, 11, 30)
-            },
-        ],
-        views: [{
-            type: "week",
-            groupOrientation: "vertical"
-        }],
-        currentView: "week",
-        groups: ["id"],
-        resources: [
-            {
-                field: "id",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        currentDate: new Date(2018, 4, 21),
-        startDayHour: 9,
-        endDayHour: 15,
-        cellDuration: 60,
-        showAllDayPanel: true,
-        maxAppointmentsPerCell: 'auto'
-    });
-
-    var $appointments = $(this.instance.$element()).find("." + APPOINTMENT_CLASS);
-    assert.equal($appointments.length, 2, "two appointments are rendered");
-
-    var cellHeight = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).first().outerHeight();
-
-    assert.roughEqual($appointments.eq(0).position().top, 0, 1.5, "correct top position of allDay appointment");
-    assert.roughEqual($appointments.eq(0).outerHeight(), 0.5 * cellHeight, 2, "correct size of allDay appointment");
-    assert.roughEqual($appointments.eq(0).position().left, 314, 1, "correct left position of allDay appointment");
-
-    assert.roughEqual($appointments.eq(1).position().top, 7 * cellHeight, 1.5, "correct top position of allDay appointment");
-    assert.roughEqual($appointments.eq(1).outerHeight(), 0.5 * cellHeight, 2, "correct size of allDay appointment");
-    assert.roughEqual($appointments.eq(1).position().left, 314, 1, "correct left position of allDay appointment");
-});
-
-QUnit.test("Rival allDay appointments from same groups should be rendered correctly in vertical grouped workspace Week", function(assert) {
-    this.createInstance({
-        dataSource: [
-            {
-                text: "1",
-                id: 1,
-                allDay: true,
-                startDate: new Date(2018, 4, 21, 9, 30),
-                endDate: new Date(2018, 4, 21, 11, 30)
-            },
-            {
-                text: "2",
-                id: 1,
-                allDay: true,
-                startDate: new Date(2018, 4, 21, 9, 30),
-                endDate: new Date(2018, 4, 21, 11, 30)
-            },
-        ],
-        views: [{
-            type: "week",
-            groupOrientation: "vertical"
-        }],
-        currentView: "week",
-        groups: ["id"],
-        resources: [
-            {
-                field: "id",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        currentDate: new Date(2018, 4, 21),
-        startDayHour: 9,
-        endDayHour: 15,
-        cellDuration: 60,
-        showAllDayPanel: true,
-        maxAppointmentsPerCell: 'auto'
-    });
-
-    var $appointments = $(this.instance.$element()).find("." + APPOINTMENT_CLASS);
-
-    var cellHeight = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).first().outerHeight();
-
-    assert.roughEqual($appointments.eq(0).position().top, 0.5 * cellHeight, 2.5, "correct top position of allDay appointment");
-    assert.roughEqual($appointments.eq(0).outerHeight(), 0.5 * cellHeight, 2, "correct size of allDay appointment");
-    assert.roughEqual($appointments.eq(0).position().left, 314, 1, "correct left position of allDay appointment");
-});
-
-QUnit.test("Rival appointments from one group should be rendered correctly in vertical grouped workspace Week", function(assert) {
-    this.createInstance({
-        dataSource: [],
-        views: [{
-            type: "week",
-            groupOrientation: "vertical"
-        }],
-        currentView: "week",
-        groups: ["id"],
-        resources: [
-            {
-                field: "id",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        currentDate: new Date(2018, 4, 21),
-        startDayHour: 9,
-        endDayHour: 15,
-        cellDuration: 60,
-        showAllDayPanel: true,
-        maxAppointmentsPerCell: null
-    });
-
-    var defaultWidthStub = sinon.stub(this.instance.getRenderingStrategyInstance(), "_getAppointmentMaxWidth").returns(50);
-
-    this.instance.option("dataSource", [
-        {
-            text: "1",
-            id: 2,
-            startDate: new Date(2018, 4, 21, 9, 30),
-            endDate: new Date(2018, 4, 21, 11, 30)
-        },
-        {
-            text: "2",
-            id: 2,
-            startDate: new Date(2018, 4, 22, 9, 30),
-            endDate: new Date(2018, 4, 22, 11, 30)
-        },
-    ]);
-
-    var $appointments = $(this.instance.$element()).find("." + APPOINTMENT_CLASS);
-    assert.equal($appointments.length, 2, "two appointments are rendered");
-
-    var cellHeight = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).first().outerHeight();
-
-    assert.roughEqual($appointments.eq(0).position().top, 8.5 * cellHeight, 1.5, "correct top position of appointment");
-    assert.roughEqual($appointments.eq(0).outerWidth(), 50, 2, "correct size of appointment");
-    assert.equal($appointments.eq(0).position().left, 314, "correct left position of appointment");
-
-    assert.roughEqual($appointments.eq(1).position().top, 8.5 * cellHeight, 1.5, "correct top position of appointment");
-    assert.roughEqual($appointments.eq(1).outerWidth(), 50, 2, "correct size of appointment");
-    assert.equal($appointments.eq(1).position().left, 428, "correct left position of appointment");
-
-    defaultWidthStub.restore();
-});
-
-QUnit.test("Appointment in bottom cell should be rendered cirrectly in vertical grouped workspace Week", function(assert) {
-    this.createInstance({
-        dataSource: [
-            {
-                text: "1",
-                id: 2,
-                startDate: new Date(2018, 4, 22, 13, 0),
-                endDate: new Date(2018, 4, 22, 17, 30)
-            },
-        ],
-        views: [{
-            type: "week",
-            groupOrientation: "vertical"
-        }],
-        currentView: "week",
-        groups: ["id"],
-        resources: [
-            {
-                field: "id",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        currentDate: new Date(2018, 4, 21),
-        startDayHour: 9,
-        endDayHour: 15,
-        width: 2000,
-        cellDuration: 60,
-        showAllDayPanel: true,
-        maxAppointmentsPerCell: 'auto'
-    });
-
-    var $appointments = $(this.instance.$element()).find("." + APPOINTMENT_CLASS);
-
-    var cellHeight = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).first().outerHeight();
-
-    assert.roughEqual($appointments.eq(0).position().top, 12 * cellHeight, 1.5, "correct top position of appointment");
-    assert.roughEqual($appointments.eq(0).outerHeight(), 100, 2, "correct size of appointment");
 });
 
 QUnit.test("Appointments should be rendered correctly in vertical grouped workspace Month", function(assert) {
@@ -6124,94 +3890,6 @@ QUnit.test("Appointments should be rendered correctly in vertical grouped worksp
     assert.roughEqual($appointments.eq(0).position().left, cellPosition, 1.5, "correct left position");
     assert.roughEqual($appointments.eq(1).position().top, cellHeight * 8 + monthTopOffset, 3.5, "correct top position");
     assert.roughEqual($appointments.eq(1).position().left, cellPosition, 1.5, "correct left position");
-});
-
-QUnit.test("Appointment should be dragged correctly between the groups in vertical grouped workspace Day", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            text: "a",
-            startDate: new Date(2018, 2, 1, 12),
-            endDate: new Date(2018, 2, 1, 12, 30),
-            id: 1
-        }],
-        currentDate: new Date(2018, 2, 1),
-        views: [{
-            type: "day",
-            groupOrientation: "vertical"
-        }],
-        editing: true,
-        currentView: "day",
-        groups: ["id"],
-        resources: [
-            {
-                field: "id",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        startDayHour: 12,
-        endDayHour: 16,
-        showAllDayPanel: false
-    });
-
-    var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(0);
-
-    $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).eq(10).trigger(dragEvents.enter);
-    $appointment.trigger(dragEvents.start);
-    $appointment.trigger(dragEvents.end);
-
-    this.clock.tick();
-    var appointmentData = dataUtils.data(this.instance.$element().find("." + APPOINTMENT_CLASS).get(0), "dxItemData");
-
-    assert.deepEqual(appointmentData.startDate, new Date(2018, 2, 1, 13), "Start date is correct");
-    assert.deepEqual(appointmentData.endDate, new Date(2018, 2, 1, 13, 30), "End date is correct");
-    assert.deepEqual(appointmentData.id, 2, "Group is OK");
-});
-
-QUnit.test("Appointment should be dragged correctly between the groups in vertical grouped workspace Week", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            text: "a",
-            startDate: new Date(2018, 2, 16, 12),
-            endDate: new Date(2018, 2, 16, 12, 30),
-            id: 1
-        }],
-        currentDate: new Date(2018, 2, 16),
-        views: [{
-            type: "week",
-            groupOrientation: "vertical"
-        }],
-        editing: true,
-        currentView: "week",
-        groups: ["id"],
-        resources: [
-            {
-                field: "id",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        startDayHour: 12,
-        endDayHour: 16,
-        showAllDayPanel: false
-    });
-
-    var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).eq(0);
-
-    $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).eq(75).trigger(dragEvents.enter);
-    $appointment.trigger(dragEvents.start);
-    $appointment.trigger(dragEvents.end);
-
-    this.clock.tick();
-    var appointmentData = dataUtils.data(this.instance.$element().find("." + APPOINTMENT_CLASS).get(0), "dxItemData");
-
-    assert.deepEqual(appointmentData.startDate, new Date(2018, 2, 16, 13), "Start date is correct");
-    assert.deepEqual(appointmentData.endDate, new Date(2018, 2, 16, 13, 30), "End date is correct");
-    assert.deepEqual(appointmentData.id, 2, "Group is OK");
 });
 
 QUnit.test("Appointment should be dragged correctly between the groups in vertical grouped workspace Month", function(assert) {
@@ -6331,41 +4009,6 @@ QUnit.test("Long appt parts should have correct coordinates after drag to the la
     assert.equal($secondPart.position().left, 0, 'correct left position');
 });
 
-QUnit.test("Hourly recurring appt should be rendred in vertical grouped workspace Day", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            text: "a",
-            startDate: new Date(2018, 2, 16, 12, 30),
-            endDate: new Date(2018, 2, 16, 13, 15),
-            recurrenceRule: "FREQ=HOURLY",
-            id: 1
-        }],
-        currentDate: new Date(2018, 2, 16),
-        views: [{
-            type: "day",
-            groupOrientation: "vertical"
-        }],
-        currentView: "day",
-        groups: ["id"],
-        resources: [
-            {
-                field: "id",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        startDayHour: 12,
-        endDayHour: 16,
-        showAllDayPanel: false
-    });
-
-    var $appointments = $(this.instance.$element()).find("." + APPOINTMENT_CLASS);
-
-    assert.equal($appointments.length, 4, "Appointments are rendered");
-});
-
 QUnit.test("Appointment should be resized correctly to left side in horizontal grouped workspace Month", function(assert) {
     this.createInstance({
         dataSource: [{
@@ -6405,156 +4048,6 @@ QUnit.test("Appointment should be resized correctly to left side in horizontal g
     assert.roughEqual($appointment.position().left, 0, 1.1, "Left coordinate is correct");
 });
 
-QUnit.test("Appt shouldn't be resized to the group border in horizontal grouped workspace Day", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            text: "a",
-            startDate: new Date(2018, 2, 16, 14),
-            endDate: new Date(2018, 2, 16, 15),
-            id: 1
-        }],
-        currentDate: new Date(2018, 2, 16),
-        views: [{
-            type: "day",
-            groupOrientation: "vertical"
-        }],
-        currentView: "day",
-        groups: ["id"],
-        resources: [
-            {
-                field: "id",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        editing: true,
-        startDayHour: 12,
-        endDayHour: 16,
-        showAllDayPanel: false
-    });
-
-    var $element = $(this.instance.$element()),
-        cellHeight = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).eq(0).get(0).getBoundingClientRect().height,
-        pointer = pointerMock($element.find(".dx-resizable-handle-bottom").eq(0)).start();
-
-    pointer.dragStart().drag(0, cellHeight * 2).dragEnd();
-
-    var $appointment = $element.find("." + APPOINTMENT_CLASS).eq(0);
-
-    assert.equal($appointment.position().top + $appointment.outerHeight(), cellHeight * 8, "Correct bottom coordinate");
-});
-
-QUnit.test("Appt shouldn't be resized to the group border after scrolling in horizontal grouped workspace Day", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            text: "a",
-            startDate: new Date(2018, 2, 16, 14),
-            endDate: new Date(2018, 2, 16, 15),
-            id: 2
-        }],
-        currentDate: new Date(2018, 2, 16),
-        views: [{
-            type: "day",
-            groupOrientation: "vertical"
-        }],
-        currentView: "day",
-        groups: ["id"],
-        resources: [
-            {
-                field: "id",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        editing: true,
-        startDayHour: 12,
-        endDayHour: 16,
-        showAllDayPanel: true,
-        height: 500
-    });
-
-    var scrollable = this.instance.getWorkSpace().getScrollable();
-    scrollable.scrollTo({ left: 0, top: 400 });
-
-    var $element = $(this.instance.$element()),
-        cellHeight = $(this.instance.$element()).find("." + DATE_TABLE_CELL_CLASS).eq(0).outerHeight(),
-        pointer = pointerMock($element.find(".dx-resizable-handle-top").eq(0)).start();
-
-    pointer.dragStart().drag(0, -cellHeight * 5).dragEnd();
-
-    var $appointment = $element.find("." + APPOINTMENT_CLASS).eq(0);
-
-    assert.roughEqual($appointment.position().top, 500, 1.1, "Correct top coordinate");
-    assert.equal($appointment.outerHeight(), cellHeight * 6, "Correct height");
-});
-
-QUnit.test("Appointment inside vertical grouped view should have a right resizable area in Day view", function(assert) {
-    this.createInstance({
-        dataSource: [{
-            text: "a",
-            startDate: new Date(2018, 2, 16, 14),
-            endDate: new Date(2018, 2, 16, 15),
-            id: 1
-        }],
-        currentDate: new Date(2018, 2, 16),
-        views: [{
-            type: "day",
-            groupOrientation: "vertical"
-        }],
-        currentView: "day",
-        groups: ["id"],
-        editing: true,
-        resources: [
-            {
-                field: "id",
-                dataSource: [
-                    { id: 1, text: "one" },
-                    { id: 2, text: "two" }
-                ]
-            }
-        ],
-        startDayHour: 12,
-        endDayHour: 16,
-        showAllDayPanel: false
-    });
-
-    var $appointment = $(this.instance.$element()).find("." + APPOINTMENT_CLASS).first(),
-        initialResizableAreaTop = $appointment.dxResizable("instance").option("area").top,
-        initialResizableAreaBottom = $appointment.dxResizable("instance").option("area").bottom;
-
-    assert.equal($appointment.dxResizable("instance").option("area").top, initialResizableAreaTop);
-    assert.equal($appointment.dxResizable("instance").option("area").bottom, initialResizableAreaBottom);
-});
-
-QUnit.test("New added appointment should be rendered correctly in specified timeZone", function(assert) {
-    var tzOffsetStub = sinon.stub(subscribes, "getClientTimezoneOffset").returns(-10800000);
-    try {
-        this.createInstance({
-            dataSource: [],
-            currentDate: new Date(2018, 4, 25),
-            views: ["week"],
-            currentView: "week",
-            timeZone: "Etc/UTC"
-        });
-
-        var task = { text: "a", startDate: new Date(2018, 4, 23, 8, 0), endDate: new Date(2018, 4, 23, 8, 30) };
-
-        this.instance.showAppointmentPopup(task, true);
-        $(".dx-scheduler-appointment-popup .dx-popup-done").trigger("dxclick");
-
-        var $appointment = this.instance.$element().find("." + APPOINTMENT_CLASS),
-            startDate = $appointment.dxSchedulerAppointment("instance").option("startDate");
-
-        assert.deepEqual(startDate, task.startDate, "appointment starts in 8AM");
-    } finally {
-        tzOffsetStub.restore();
-    }
-});
-
 QUnit.test("Tail of long appointment should have a right position, groupByDate = true", function(assert) {
     this.createInstance({
         dataSource: [
@@ -6591,3 +4084,4 @@ QUnit.test("Tail of long appointment should have a right position, groupByDate =
 
     assert.roughEqual($appointmentTail.position().left, $cell.position().left, 1.001, "Tail has a right position");
 });
+
