+++ testing/tests/DevExpress.ui.widgets.scheduler/workSpace.tests.js
@@ -952,6 +952,37 @@ QUnit.testStart(function() {
         assert.deepEqual(secondCellData.startDate, new Date(2018, 2, 14, 11, 30), "cell has right startDate");
         assert.deepEqual(secondCellData.endDate, new Date(2018, 2, 14, 12), "cell has right endtDate");
     });
+
+    QUnit.test("Vertical grouped work space should calculate max top position", function(assert) {
+        this.instance.option({
+            currentDate: new Date(2015, 2, 16),
+            firstDayOfWeek: 1,
+            showAllDayPanel: true,
+            startDayHour: 8,
+            endDayHour: 9,
+            groupOrientation: "vertical",
+            groups: [{
+                name: "one",
+                items: [{ id: 1, text: "a" }, { id: 2, text: "b" }]
+            },
+            {
+                name: "two",
+                items: [{ id: 1, text: "c" }, { id: 2, text: "d" }]
+            }]
+        });
+
+        var $rows = this.instance.$element().find(".dx-scheduler-date-table tr"),
+            $firstGroupLastCell = $rows.eq(2).find("td").first(),
+            $secondGroupLastCell = $rows.eq(5).find("td").first(),
+            $thirdGroupLastCell = $rows.eq(8).find("td").first(),
+            $fourthGroupLastCell = $rows.eq(11).find("td").first();
+
+        assert.equal($firstGroupLastCell.position().top + $firstGroupLastCell.get(0).getBoundingClientRect().height, this.instance.getVerticalMax(0), "Max top is OK");
+        assert.equal($secondGroupLastCell.position().top + $secondGroupLastCell.get(0).getBoundingClientRect().height, this.instance.getVerticalMax(1), "Max top is OK");
+        assert.equal($thirdGroupLastCell.position().top + $thirdGroupLastCell.get(0).getBoundingClientRect().height, this.instance.getVerticalMax(2), "Max top is OK");
+        assert.equal($fourthGroupLastCell.position().top + $fourthGroupLastCell.get(0).getBoundingClientRect().height, this.instance.getVerticalMax(3), "Max top is OK");
+    });
+
 })("Work Space Week");
